chore:封版 * chore:封版 * chore:关闭自动更新 * new:同步职业平衡
