+++ Characters/极诣·鬼泣.py
@@ -250,7 +250,7 @@ class 技能13(主动技能):
     TP成长 = 0.10
     TP上限 = 5
     演出时间 = 5
-    data1 = [int(i*1.085) for i in [0, 846, 932, 1019, 1105, 1190, 1276, 1362, 1448, 1534, 1620, 1705, 1791, 1877, 1964, 2050, 2135, 2221, 2307, 2393, 2479, 2565, 2650, 2736, 2822, 2909, 2995, 3081, 3166, 3252, 3338, 3424, 3510, 3596, 3681, 3767, 3854, 3940, 4026, 4111, 4197, 4283, 4369, 4455, 4541, 4626, 4712, 4799, 4885, 4971, 5057, 5142, 5228, 5314, 5400, 5486, 5571, 5657, 5744, 5830, 5916]]
+    data1 = [int(i*1.085*1.051) for i in [0, 846, 932, 1019, 1105, 1190, 1276, 1362, 1448, 1534, 1620, 1705, 1791, 1877, 1964, 2050, 2135, 2221, 2307, 2393, 2479, 2565, 2650, 2736, 2822, 2909, 2995, 3081, 3166, 3252, 3338, 3424, 3510, 3596, 3681, 3767, 3854, 3940, 4026, 4111, 4197, 4283, 4369, 4455, 4541, 4626, 4712, 4799, 4885, 4971, 5057, 5142, 5228, 5314, 5400, 5486, 5571, 5657, 5744, 5830, 5916]]
 
     数据 = [data1]
     次数 = [10]
@@ -309,7 +309,7 @@ class 技能14(主动技能):
     TP成长 = 0.10
     TP上限 = 5
     演出时间 = 0.5
-    data1 = [int(i*1.085) for i in [0, 9609, 10584, 11558, 12533, 13508, 14483, 15458, 16432, 17407, 18382, 19357, 20333, 21307, 22282, 23257, 24232, 25207, 26181, 27156, 28131, 29106, 30081, 31056, 32030, 33005, 33980, 34955, 35930, 36904, 37879, 38854, 39829, 40805, 41778, 42754, 43729, 44704, 45679, 46653, 47628, 48603, 49578, 50553, 51527, 52502, 53477, 54452, 55427, 56402, 57376, 58351, 59326, 60301, 61276, 62250, 63226, 64201, 65176, 66151, 67125, 68100, 69075, 70050, 71025, 71999, 72974, 73949, 74924, 75899, 76874]]
+    data1 = [int(i*1.085*1.052) for i in [0, 9609, 10584, 11558, 12533, 13508, 14483, 15458, 16432, 17407, 18382, 19357, 20333, 21307, 22282, 23257, 24232, 25207, 26181, 27156, 28131, 29106, 30081, 31056, 32030, 33005, 33980, 34955, 35930, 36904, 37879, 38854, 39829, 40805, 41778, 42754, 43729, 44704, 45679, 46653, 47628, 48603, 49578, 50553, 51527, 52502, 53477, 54452, 55427, 56402, 57376, 58351, 59326, 60301, 61276, 62250, 63226, 64201, 65176, 66151, 67125, 68100, 69075, 70050, 71025, 71999, 72974, 73949, 74924, 75899, 76874]]
 
     数据 = [data1]
     次数 = [1]
@@ -421,10 +421,10 @@ class 技能18(主动技能):
     基础等级 = 1
     CD = 45.0
     TP成长 = 0.10
-    data1 = [int(i*1.085) for i in [0, 943, 1038, 1134, 1231, 1326, 1422, 1517, 1613, 1708, 1804, 1901, 1996, 2092, 2187, 2283, 2378, 2475, 2570, 2666, 2762, 2857, 2953, 3048, 3145, 3240, 3336, 3431, 3527, 3623, 3719, 3815, 3910, 4006, 4101, 4197, 4292, 4389, 4485, 4580, 4676, 4771, 4867, 4963, 5059, 5154, 5250, 5346, 5441, 5538, 5633, 5729, 5824, 5920, 6015, 6111, 6208, 6303, 6399, 6494, 6590, 6685, 6782, 6878, 6973, 7069, 7164, 7260, 7355, 7452, 7547]]
-    data2 = [int(i*1.085) for i in [0, 1381, 1520, 1661, 1801, 1941, 2081, 2222, 2361, 2502, 2642, 2783, 2922, 3063, 3203, 3342, 3483, 3623, 3763, 3903, 4044, 4183, 4324, 4464, 4604, 4744, 4884, 5024, 5164, 5305, 5444, 5585, 5725, 5865, 6005, 6146, 6285, 6425, 6566, 6705, 6846, 6986, 7126, 7266, 7407, 7546, 7687, 7827, 7967, 8107, 8247, 8387, 8527, 8668, 8807, 8948, 9088, 9228, 9368, 9509, 9648, 9788, 9929, 10068, 10209, 10349, 10490, 10629, 10770, 10910, 11050]]
-    data3 = [int(i*1.085) for i in [0, 2961, 3261, 3563, 3863, 4163, 4464, 4764, 5064, 5366, 5666, 5966, 6266, 6567, 6867, 7168, 7469, 7769, 8069, 8370, 8670, 8971, 9272, 9572, 9872, 10172, 10473, 10774, 11074, 11375, 11675, 11975, 12276, 12577, 12877, 13178, 13478, 13778, 14078, 14380, 14680, 14980, 15281, 15581, 15881, 16183, 16483, 16783, 17084, 17384, 17684, 17985, 18286, 18586, 18886, 19187, 19487, 19788, 20089, 20389, 20689, 20990, 21290, 21591, 21891, 22192, 22492, 22792, 23093, 23394, 23694]]
-    data4 = [int(i*1.085) for i in [0, 4810, 5298, 5786, 6274, 6762, 7250, 7737, 8226, 8714, 9202, 9690, 10178, 10666, 11154, 11642, 12130, 12618, 13106, 13594, 14082, 14570, 15058, 15546, 16035, 16522, 17010, 17498, 17987, 18474, 18962, 19450, 19938, 20427, 20914, 21402, 21890, 22379, 22866, 23354, 23842, 24331, 24818, 25306, 25794, 26283, 26771, 27258, 27746, 28235, 28723, 29210, 29698, 30186, 30675, 31163, 31650, 32138, 32627, 33115, 33602, 34090, 34579, 35067, 35554, 36042, 36531, 37019, 37507, 37994, 38483]]
+    data1 = [int(i*1.0858*1.053) for i in [0, 943, 1038, 1134, 1231, 1326, 1422, 1517, 1613, 1708, 1804, 1901, 1996, 2092, 2187, 2283, 2378, 2475, 2570, 2666, 2762, 2857, 2953, 3048, 3145, 3240, 3336, 3431, 3527, 3623, 3719, 3815, 3910, 4006, 4101, 4197, 4292, 4389, 4485, 4580, 4676, 4771, 4867, 4963, 5059, 5154, 5250, 5346, 5441, 5538, 5633, 5729, 5824, 5920, 6015, 6111, 6208, 6303, 6399, 6494, 6590, 6685, 6782, 6878, 6973, 7069, 7164, 7260, 7355, 7452, 7547]]
+    data2 = [int(i*1.085*1.053) for i in [0, 1381, 1520, 1661, 1801, 1941, 2081, 2222, 2361, 2502, 2642, 2783, 2922, 3063, 3203, 3342, 3483, 3623, 3763, 3903, 4044, 4183, 4324, 4464, 4604, 4744, 4884, 5024, 5164, 5305, 5444, 5585, 5725, 5865, 6005, 6146, 6285, 6425, 6566, 6705, 6846, 6986, 7126, 7266, 7407, 7546, 7687, 7827, 7967, 8107, 8247, 8387, 8527, 8668, 8807, 8948, 9088, 9228, 9368, 9509, 9648, 9788, 9929, 10068, 10209, 10349, 10490, 10629, 10770, 10910, 11050]]
+    data3 = [int(i*1.085*1.053) for i in [0, 2961, 3261, 3563, 3863, 4163, 4464, 4764, 5064, 5366, 5666, 5966, 6266, 6567, 6867, 7168, 7469, 7769, 8069, 8370, 8670, 8971, 9272, 9572, 9872, 10172, 10473, 10774, 11074, 11375, 11675, 11975, 12276, 12577, 12877, 13178, 13478, 13778, 14078, 14380, 14680, 14980, 15281, 15581, 15881, 16183, 16483, 16783, 17084, 17384, 17684, 17985, 18286, 18586, 18886, 19187, 19487, 19788, 20089, 20389, 20689, 20990, 21290, 21591, 21891, 22192, 22492, 22792, 23093, 23394, 23694]]
+    data4 = [int(i*1.085*1.053) for i in [0, 4810, 5298, 5786, 6274, 6762, 7250, 7737, 8226, 8714, 9202, 9690, 10178, 10666, 11154, 11642, 12130, 12618, 13106, 13594, 14082, 14570, 15058, 15546, 16035, 16522, 17010, 17498, 17987, 18474, 18962, 19450, 19938, 20427, 20914, 21402, 21890, 22379, 22866, 23354, 23842, 24331, 24818, 25306, 25794, 26283, 26771, 27258, 27746, 28235, 28723, 29210, 29698, 30186, 30675, 31163, 31650, 32138, 32627, 33115, 33602, 34090, 34579, 35067, 35554, 36042, 36531, 37019, 37507, 37994, 38483]]
 
     数据 = [data1, data2, data3, data4]
     次数 = [2, 2, 2, 1]
@@ -510,9 +510,9 @@ class 技能21(主动技能):
     CD = 30.0
     TP成长 = 0.10
     TP上限 = 5
-    data1 = [int(i*1.085) for i in [0, 1287, 1418, 1549, 1679, 1810, 1940, 2072, 2202, 2332, 2463, 2594, 2725, 2855, 2986, 3116, 3247, 3378, 3508, 3639, 3770, 3901, 4031, 4161, 4292, 4423, 4554, 4684, 4815, 4946, 5076, 5207, 5337, 5468, 5599, 5730, 5860, 5990, 6122, 6252, 6383, 6513, 6644, 6775, 6905, 7036, 7166, 7298, 7428, 7559, 7689, 7819, 7951, 8081, 8212, 8342, 8474, 8604, 8735, 8865, 8995, 9127, 9257, 9388, 9518, 9650, 9780, 9910, 10041, 10171, 10303]]
-    data2 = [int(i*1.085) for i in [0, 5007, 5516, 6024, 6532, 7040, 7548, 8057, 8565, 9072, 9580, 10088, 10597, 11105, 11613, 12121, 12629, 13138, 13645, 14153, 14661, 15169, 15678, 16186, 16694, 17202, 17710, 18219, 18726, 19234, 19742, 20250, 20759, 21267, 21775, 22283, 22791, 23298, 23807, 24315, 24823, 25331, 25839, 26348, 26856, 27364, 27871, 28379, 28888, 29396, 29904, 30412, 30920, 31429, 31937, 32444, 32952, 33460, 33969, 34477, 34985, 35493, 36001, 36510, 37018, 37525, 38033, 38541, 39050, 39558, 40066]]
-    data3 = [int(i*1.085) for i in [0, 4292, 4728, 5163, 5599, 6034, 6470, 6905, 7341, 7776, 8212, 8648, 9083, 9518, 9954, 10390, 10824, 11260, 11696, 12132, 12566, 13002, 13438, 13874, 14308, 14744, 15180, 15616, 16051, 16486, 16922, 17358, 17793, 18228, 18664, 19100, 19535, 19970, 20406, 20841, 21277, 21712, 22148, 22583, 23019, 23455, 23890, 24325, 24761, 25197, 25632, 26067, 26503, 26939, 27374, 27809, 28245, 28681, 29117, 29551, 29987, 30423, 30858, 31293, 31729, 32165, 32600, 33035, 33471, 33907, 34342]]
+    data1 = [int(i*1.085*1.05) for i in [0, 1287, 1418, 1549, 1679, 1810, 1940, 2072, 2202, 2332, 2463, 2594, 2725, 2855, 2986, 3116, 3247, 3378, 3508, 3639, 3770, 3901, 4031, 4161, 4292, 4423, 4554, 4684, 4815, 4946, 5076, 5207, 5337, 5468, 5599, 5730, 5860, 5990, 6122, 6252, 6383, 6513, 6644, 6775, 6905, 7036, 7166, 7298, 7428, 7559, 7689, 7819, 7951, 8081, 8212, 8342, 8474, 8604, 8735, 8865, 8995, 9127, 9257, 9388, 9518, 9650, 9780, 9910, 10041, 10171, 10303]]
+    data2 = [int(i*1.085*1.05) for i in [0, 5007, 5516, 6024, 6532, 7040, 7548, 8057, 8565, 9072, 9580, 10088, 10597, 11105, 11613, 12121, 12629, 13138, 13645, 14153, 14661, 15169, 15678, 16186, 16694, 17202, 17710, 18219, 18726, 19234, 19742, 20250, 20759, 21267, 21775, 22283, 22791, 23298, 23807, 24315, 24823, 25331, 25839, 26348, 26856, 27364, 27871, 28379, 28888, 29396, 29904, 30412, 30920, 31429, 31937, 32444, 32952, 33460, 33969, 34477, 34985, 35493, 36001, 36510, 37018, 37525, 38033, 38541, 39050, 39558, 40066]]
+    data3 = [int(i*1.085*1.05) for i in [0, 4292, 4728, 5163, 5599, 6034, 6470, 6905, 7341, 7776, 8212, 8648, 9083, 9518, 9954, 10390, 10824, 11260, 11696, 12132, 12566, 13002, 13438, 13874, 14308, 14744, 15180, 15616, 16051, 16486, 16922, 17358, 17793, 18228, 18664, 19100, 19535, 19970, 20406, 20841, 21277, 21712, 22148, 22583, 23019, 23455, 23890, 24325, 24761, 25197, 25632, 26067, 26503, 26939, 27374, 27809, 28245, 28681, 29117, 29551, 29987, 30423, 30858, 31293, 31729, 32165, 32600, 33035, 33471, 33907, 34342]]
 
     数据 = [data1, data2, data3]
     次数 = [1, 2, 1]
@@ -570,7 +570,7 @@ class 技能22(主动技能):
     CD = 40.0
     TP成长 = 0.10
     TP上限 = 5
-    data1 = [int(i*1.085) for i in [0, 16391, 18055, 19718, 21380, 23044, 24707, 26371, 28033, 29696, 31358, 33022, 34684, 36347, 38011, 39673, 41336, 43000, 44663, 46326, 47989, 49651, 51315, 52978, 54640, 56304, 57966, 59629, 61293, 62956, 64618, 66282, 67945, 69607, 71271, 72933, 74596,
+    data1 = [int(i*1.085*1.052) for i in [0, 16391, 18055, 19718, 21380, 23044, 24707, 26371, 28033, 29696, 31358, 33022, 34684, 36347, 38011, 39673, 41336, 43000, 44663, 46326, 47989, 49651, 51315, 52978, 54640, 56304, 57966, 59629, 61293, 62956, 64618, 66282, 67945, 69607, 71271, 72933, 74596,
              76259, 77922, 79586, 81249, 82912, 84575, 86238, 87900, 89564, 91226, 92889, 94551, 96215, 97878, 99542, 101205, 102867, 104531, 106193, 107856, 109519, 111182, 112846, 114508, 116171, 117835, 119498, 121160, 122824, 124486, 126149, 127813, 129475, 131138]]
     数据 = [data1]
     次数 = [1]
@@ -631,9 +631,9 @@ class 技能24(主动技能):
     等级上限 = 40
     基础等级 = 16
     CD = 40.0
-    data1 = [int(i*1.085) for i in [0, 330, 364, 397, 431, 464, 499, 532, 566, 599, 632, 666, 699, 733, 767, 801, 834, 868, 901, 934, 968, 1001, 1036, 1069, 1103, 1136, 1170, 1203, 1237, 1270, 1303, 1338, 1371, 1405, 1438, 1472, 1505, 1539, 1572, 1606, 1640, 1673, 1707, 1740, 1774, 1807, 1841, 1875, 1909, 1942, 1975, 2009, 2042, 2076, 2109, 2143, 2177, 2211, 2244, 2277, 2311, 2344, 2378, 2411, 2446, 2479, 2513, 2546, 2580, 2613, 2646]]
-    data2 = [int(i*1.085) for i in [0, 3683, 4056, 4430, 4803, 5177, 5551, 5925, 6298, 6671, 7045, 7419, 7793, 8166, 8540, 8913, 9287, 9661, 10035, 10408, 10782, 11155, 11530, 11903, 12276, 12650, 13023, 13398, 13771, 14145, 14518, 14892, 15265, 15640, 16013, 16386, 16760, 17133, 17508, 17881, 18255, 18628, 19002, 19375, 19750, 20123, 20497, 20870, 21243, 21618, 21991, 22365, 22738, 23112, 23485, 23860, 24233, 24607, 24980, 25354, 25728, 26101, 26475, 26848, 27222, 27595, 27970, 28343, 28717, 29090, 29464]]
-    data3 = [int(i*1.085) for i in [0, 20717, 22818, 24920, 27022, 29123, 31225, 33327, 35429, 37531, 39633, 41734, 43836, 45938, 48039, 50141, 52243, 54344, 56446, 58549, 60650, 62752, 64854, 66955, 69057, 71159, 73260, 75362, 77464, 79565, 81668, 83770, 85871, 87973, 90074, 92176, 94278, 96379, 98481, 100583, 102684, 104787, 106889, 108990, 111092, 113194, 115295, 117397, 119499, 121600, 123702, 125805, 127906, 130008, 132110, 134211, 136313, 138415, 140516, 142618, 144720, 146821, 148924, 151026, 153127, 155229, 157331, 159432, 161534, 163635, 165737]]
+    data1 = [int(i*1.085*1.051) for i in [0, 330, 364, 397, 431, 464, 499, 532, 566, 599, 632, 666, 699, 733, 767, 801, 834, 868, 901, 934, 968, 1001, 1036, 1069, 1103, 1136, 1170, 1203, 1237, 1270, 1303, 1338, 1371, 1405, 1438, 1472, 1505, 1539, 1572, 1606, 1640, 1673, 1707, 1740, 1774, 1807, 1841, 1875, 1909, 1942, 1975, 2009, 2042, 2076, 2109, 2143, 2177, 2211, 2244, 2277, 2311, 2344, 2378, 2411, 2446, 2479, 2513, 2546, 2580, 2613, 2646]]
+    data2 = [int(i*1.085*1.051) for i in [0, 3683, 4056, 4430, 4803, 5177, 5551, 5925, 6298, 6671, 7045, 7419, 7793, 8166, 8540, 8913, 9287, 9661, 10035, 10408, 10782, 11155, 11530, 11903, 12276, 12650, 13023, 13398, 13771, 14145, 14518, 14892, 15265, 15640, 16013, 16386, 16760, 17133, 17508, 17881, 18255, 18628, 19002, 19375, 19750, 20123, 20497, 20870, 21243, 21618, 21991, 22365, 22738, 23112, 23485, 23860, 24233, 24607, 24980, 25354, 25728, 26101, 26475, 26848, 27222, 27595, 27970, 28343, 28717, 29090, 29464]]
+    data3 = [int(i*1.085*1.051) for i in [0, 20717, 22818, 24920, 27022, 29123, 31225, 33327, 35429, 37531, 39633, 41734, 43836, 45938, 48039, 50141, 52243, 54344, 56446, 58549, 60650, 62752, 64854, 66955, 69057, 71159, 73260, 75362, 77464, 79565, 81668, 83770, 85871, 87973, 90074, 92176, 94278, 96379, 98481, 100583, 102684, 104787, 106889, 108990, 111092, 113194, 115295, 117397, 119499, 121600, 123702, 125805, 127906, 130008, 132110, 134211, 136313, 138415, 140516, 142618, 144720, 146821, 148924, 151026, 153127, 155229, 157331, 159432, 161534, 163635, 165737]]
     数据 = [data1, data2, data3]
     次数 = [8 * 4, 4, 1]
     解放 = 1
@@ -659,8 +659,8 @@ class 技能25(主动技能):
     等级上限 = 40
     基础等级 = 13
     CD = 45.0
-    data1 = [int(i*1.085) for i in [0, 18254, 20105, 21958, 23810, 25661, 27513, 29365, 31218, 33069, 34921, 36773, 38624, 40476, 42329, 44181, 46032, 47884, 49736, 51588, 53440, 55292, 57144, 58995, 60848, 62700, 64551, 66403, 68255, 70107, 71959, 73811, 75663, 77514, 79366, 81219, 83071, 84922, 86774, 88626, 90477]]
-    data2 = [int(i*1.085) for i in [0, 27381, 30159, 32936, 35715, 38492, 41270, 44049, 46826, 49604, 52381, 55160, 57937, 60715, 63494, 66271, 69049, 71826, 74605, 77382, 80160, 82938, 85716, 88494, 91271, 94050, 96827, 99605, 102382, 105161, 107939, 110716, 113495, 116272, 119050, 121827, 124606, 127384, 130161, 132939, 135717]]
+    data1 = [int(i*1.085*1.05) for i in [0, 18254, 20105, 21958, 23810, 25661, 27513, 29365, 31218, 33069, 34921, 36773, 38624, 40476, 42329, 44181, 46032, 47884, 49736, 51588, 53440, 55292, 57144, 58995, 60848, 62700, 64551, 66403, 68255, 70107, 71959, 73811, 75663, 77514, 79366, 81219, 83071, 84922, 86774, 88626, 90477]]
+    data2 = [int(i*1.085*1.05) for i in [0, 27381, 30159, 32936, 35715, 38492, 41270, 44049, 46826, 49604, 52381, 55160, 57937, 60715, 63494, 66271, 69049, 71826, 74605, 77382, 80160, 82938, 85716, 88494, 91271, 94050, 96827, 99605, 102382, 105161, 107939, 110716, 113495, 116272, 119050, 121827, 124606, 127384, 130161, 132939, 135717]]
     数据 = [data1, data2]
     次数 = [1, 1]
 
@@ -684,10 +684,10 @@ class 技能26(主动技能):
     等级上限 = 40
     基础等级 = 5
     CD = 180.0
-    data1 = [int(i*1.085) for i in [0, 16629, 20484, 24340, 28197, 32052, 35909, 39765, 43620, 47477, 51332, 55188, 59045, 62900, 66757, 70613, 74468, 78325, 82180, 86036, 89893, 93748, 97605, 101460, 105316, 109173, 113028, 116885, 120741, 124596, 128453, 132308, 136164, 140021, 143876, 147733, 151589, 155444, 159301, 163156, 167012, 170869, 174724, 178581, 182436, 186292, 190149, 194004, 197860, 201717, 205572, 209429, 213284, 217140, 220997, 224852, 228708, 232565, 236420, 240277, 244132, 247988, 251845, 255700, 259557, 263412, 267268, 271125, 274980, 278836, 282693]]
-    data2 = [int(i*1.085) for i in [0, 831, 1024, 1217, 1409, 1602, 1794, 1988, 2181, 2373, 2566, 2759, 2952, 3145, 3337, 3530, 3723, 3915, 4109, 4301, 4494, 4687, 4879, 5073, 5266, 5458, 5651, 5843, 6036, 6230, 6422, 6615, 6808, 7000, 7194, 7386, 7579, 7772, 7964, 8157, 8350, 8543, 8736, 8928, 9121, 9315, 9507, 9700, 9892, 10085, 10278, 10471, 10664, 10857, 11049, 11242, 11435, 11628, 11821, 12013, 12206, 12398, 12592, 12785, 12977, 13170, 13363, 13556, 13749, 13941, 14134]]
-    data3 = [int(i*1.085) for i in [0, 20786, 25605, 30426, 35245, 40066, 44885, 49706, 54525, 59346, 64166, 68986, 73806, 78626, 83446, 88266, 93086, 97906, 102726, 107545, 112366, 117185, 122006, 126825, 131646, 136465, 141286, 146105, 150926, 155745, 160566, 165386, 170206, 175026, 179846, 184666, 189486, 194306, 199125, 203946, 208765, 213586, 218405, 223226, 228045, 232866, 237685, 242506, 247325, 252146, 256965, 261786, 266606, 271426, 276246, 281066, 285886, 290705, 295526, 300345, 305166, 309985, 314806, 319625, 324446, 329265, 334086, 338905, 343726, 348545, 353366]]
-    data4 = [int(i*1.085) for i in [0, 33258, 40970, 48681, 56393, 64106, 71818, 79529, 87241, 94954, 102666, 110377, 118089, 125802, 133514, 141225, 148937, 156650, 164362, 172073, 179785, 187497, 195210, 202921, 210633, 218345, 226058, 233770, 241481, 249193, 256906, 264618, 272329, 280041, 287754, 295466, 303177, 310889, 318602, 326314, 334025, 341737, 349449, 357162, 364873, 372585, 380297, 388010, 395721, 403433, 411145, 418858, 426569, 434281, 441993, 449706, 457417, 465129, 472841, 480554, 488266, 495977, 503689, 511401, 519114, 526825, 534537, 542249, 549962, 557673, 565385]]
+    data1 = [int(i*1.085*1.051) for i in [0, 16629, 20484, 24340, 28197, 32052, 35909, 39765, 43620, 47477, 51332, 55188, 59045, 62900, 66757, 70613, 74468, 78325, 82180, 86036, 89893, 93748, 97605, 101460, 105316, 109173, 113028, 116885, 120741, 124596, 128453, 132308, 136164, 140021, 143876, 147733, 151589, 155444, 159301, 163156, 167012, 170869, 174724, 178581, 182436, 186292, 190149, 194004, 197860, 201717, 205572, 209429, 213284, 217140, 220997, 224852, 228708, 232565, 236420, 240277, 244132, 247988, 251845, 255700, 259557, 263412, 267268, 271125, 274980, 278836, 282693]]
+    data2 = [int(i*1.085*1.051) for i in [0, 831, 1024, 1217, 1409, 1602, 1794, 1988, 2181, 2373, 2566, 2759, 2952, 3145, 3337, 3530, 3723, 3915, 4109, 4301, 4494, 4687, 4879, 5073, 5266, 5458, 5651, 5843, 6036, 6230, 6422, 6615, 6808, 7000, 7194, 7386, 7579, 7772, 7964, 8157, 8350, 8543, 8736, 8928, 9121, 9315, 9507, 9700, 9892, 10085, 10278, 10471, 10664, 10857, 11049, 11242, 11435, 11628, 11821, 12013, 12206, 12398, 12592, 12785, 12977, 13170, 13363, 13556, 13749, 13941, 14134]]
+    data3 = [int(i*1.085*1.051) for i in [0, 20786, 25605, 30426, 35245, 40066, 44885, 49706, 54525, 59346, 64166, 68986, 73806, 78626, 83446, 88266, 93086, 97906, 102726, 107545, 112366, 117185, 122006, 126825, 131646, 136465, 141286, 146105, 150926, 155745, 160566, 165386, 170206, 175026, 179846, 184666, 189486, 194306, 199125, 203946, 208765, 213586, 218405, 223226, 228045, 232866, 237685, 242506, 247325, 252146, 256965, 261786, 266606, 271426, 276246, 281066, 285886, 290705, 295526, 300345, 305166, 309985, 314806, 319625, 324446, 329265, 334086, 338905, 343726, 348545, 353366]]
+    data4 = [int(i*1.085*1.051) for i in [0, 33258, 40970, 48681, 56393, 64106, 71818, 79529, 87241, 94954, 102666, 110377, 118089, 125802, 133514, 141225, 148937, 156650, 164362, 172073, 179785, 187497, 195210, 202921, 210633, 218345, 226058, 233770, 241481, 249193, 256906, 264618, 272329, 280041, 287754, 295466, 303177, 310889, 318602, 326314, 334025, 341737, 349449, 357162, 364873, 372585, 380297, 388010, 395721, 403433, 411145, 418858, 426569, 434281, 441993, 449706, 457417, 465129, 472841, 480554, 488266, 495977, 503689, 511401, 519114, 526825, 534537, 542249, 549962, 557673, 565385]]
 
     数据 = [data1, data2, data3, data4]
     次数 = [1, 15, 1, 1]
