+++ Characters/知源·召唤师.py
@@ -288,8 +288,8 @@ class 技能15(知源·召唤师主动技能):
     所在等级 = 50
     等级上限 = 40
     基础等级 = 12
-    秒伤基础 = 229.8*1.1426*1.06467*1.085 + 516.2666*1.1426/2*1.06467*1.085 + 1100.6666*1.1426/2*1.06467*1.085
-    秒伤成长 = 69.2*1.1426*1.06467*1.085 + 155.7333*1.1426/1.9*1.06467*1.085 + 332.3333*1.1426/2*1.06467*1.085
+    秒伤基础 = 1.051*(229.8*1.1426*1.06467*1.085 + 516.2666*1.1426/2*1.06467*1.085 + 1100.6666*1.1426/2*1.06467*1.085)
+    秒伤成长 = 1.051*(69.2*1.1426*1.06467*1.085 + 155.7333*1.1426/1.9*1.06467*1.085 + 332.3333*1.1426/2*1.06467*1.085)
     秒伤倍率 = 0.77
     演出时间 = 0
     CD = 145
@@ -309,8 +309,8 @@ class 技能16(知源·召唤师主动技能):
 class 技能17(知源·召唤师主动技能):
     名称 = '精灵召唤：融合精灵海伊轮'
     所在等级 = 60
-    等级上限 = 20
-    基础等级 = 10
+    等级上限 = 11
+    基础等级 = 1
     是否有伤害 = 0
     自定义描述 = 1
 
@@ -321,7 +321,7 @@ class 技能17(知源·召唤师主动技能):
         if self.等级 == 0:
             return 0
         else:
-            return (2 + 2*self.等级)
+            return (20 + 2*self.等级)
 
 
 class 技能18(知源·召唤师主动技能):
@@ -329,8 +329,8 @@ class 技能18(知源·召唤师主动技能):
     所在等级 = 70
     等级上限 = 40
     基础等级 = 18
-    基础 = 3945.5*4*1.051*1.085 + 10522*1.051*1.085
-    成长 = 445.5*4*1.051*1.085 + 1188*1.051*1.085
+    基础 = (3945.5*4*1.051*1.085 + 10522*1.051*1.085)*1.071
+    成长 = (445.5*4*1.051*1.085 + 1188*1.051*1.085)*1.071
     CD = 50
     TP成长 = 0.10
     TP上限 = 5
@@ -366,8 +366,8 @@ class 技能20(知源·召唤师主动技能):
     所在等级 = 75
     等级上限 = 40
     基础等级 = 16
-    基础 = 32819.33333*1.054*1.085
-    成长 = 3705.66667*1.054*1.085
+    基础 = 32819.33333*1.054*1.085*1.065
+    成长 = 3705.66667*1.054*1.085*1.065
     CD = 40
     是否有护石 = 1
     是否带护石 = 0
@@ -387,8 +387,8 @@ class 技能21(知源·召唤师主动技能):
     所在等级 = 80
     等级上限 = 40
     基础等级 = 13
-    基础 = 3828.583333*1.05*1.085
-    成长 = 432.416667*1.05*1.085
+    基础 = 3828.583333*1.05*1.085*1.074
+    成长 = 432.416667*1.05*1.085*1.074
     攻击次数 = 15.0
     演出时间 = 3
     CD = 43
@@ -405,10 +405,10 @@ class 技能22(知源·召唤师主动技能):
     所在等级 = 85
     等级上限 = 40
     基础等级 = 5
-    基础 = 1992.4*1.15*3*1.0646*1.085
-    成长 = 601.6*1.15*3*1.0646*1.085
-    秒伤基础 = 717.6*1.15/3*2*1.0646*1.085 + 192.4*1.15/3*1.0646*1.085 + 335*1.15 /3*1.0646 + 430*1.15/3*1.0646*1.085 + 238.8*1.15/6 *1.0646*1.085 + 837.4*1.15/6*1.0646*1.085
-    秒伤成长 = 216.4*1.15/3*2*1.0646*1.085 + 57.6*1.15/3 *1.0646*1.085 + 101*1.15 /3*1.0646 + 130*1.15/3*1.0646*1.085 + 72.2*1.15/6*1.0646*1.085 + 252.6*1.15/6*1.0646*1.085
+    基础 = 1992.4*1.15*3*1.0646*1.085*1.052
+    成长 = 601.6*1.15*3*1.0646*1.085*1.052
+    秒伤基础 = 1.052*(717.6*1.15/3*2*1.0646*1.085 + 192.4*1.15/3*1.0646*1.085 + 335*1.15 /3*1.0646 + 430*1.15/3*1.0646*1.085 + 238.8*1.15/6 *1.0646*1.085 + 837.4*1.15/6*1.0646*1.085)
+    秒伤成长 = 1.052*(216.4*1.15/3*2*1.0646*1.085 + 57.6*1.15/3 *1.0646*1.085 + 101*1.15 /3*1.0646 + 130*1.15/3*1.0646*1.085 + 72.2*1.15/6*1.0646*1.085 + 252.6*1.15/6*1.0646*1.085)
     演出时间 = 1.55
     额外演出时间 = 0
     秒伤倍率 = (20 - 4.500625 * 2) / (20 - 1.55 * 2)
@@ -420,8 +420,8 @@ class 技能23(知源·召唤师主动技能):
     所在等级 = 85
     等级上限 = 40
     基础等级 = 5
-    基础 = 4782.4*1.15*1.0646*1.085 + 1912.5*1.15*5*1.0646*1.085 + 2733.2*1.15*7*1.0646*1.085 + 14346*1.15*1.0646*1.085
-    成长 = 1443.7*1.15*1.0646*1.085 + 577.7*1.15*5 *1.0646*1.085 + 824.9*1.15*7*1.0646*1.085 + 4332*1.15*1.0646*1.085
+    基础 = 1.052* (4782.4*1.15*1.0646*1.085 + 1912.5*1.15*5*1.0646*1.085 + 2733.2*1.15*7*1.0646*1.085 + 14346*1.15*1.0646*1.085)
+    成长 = 1.052* (1443.7*1.15*1.0646*1.085 + 577.7*1.15*5 *1.0646*1.085 + 824.9*1.15*7*1.0646*1.085 + 4332*1.15*1.0646*1.085)
     演出时间 = 5.25
     CD = 180
 
@@ -494,11 +494,11 @@ class 技能28(知源·召唤师主动技能):
     所在等级 = 95
     等级上限 = 40
     基础等级 = 6
-    基础 = 2162*1.3*1.1844*1.085
-    成长 = 244*1.3*1.1844*1.085
+    基础 = 2162*1.3*1.1844*1.085*1.081
+    成长 = 244*1.3*1.1844*1.085*1.081
     攻击次数 = 9.0
-    基础2 = 29185.8*1.3*1.1844*1.085
-    成长2 = 3295.2*1.3*1.1844*1.085
+    基础2 = 1.081*(29185.8*1.3*1.1844*1.085)
+    成长2 = 1.081*(295.2*1.3*1.1844*1.085)
     攻击次数2 = 1.0
     CD = 58
     演出时间 = 2.3
@@ -522,8 +522,8 @@ class 技能30(知源·召唤师主动技能):
     所在等级 = 100
     等级上限 = 40
     基础等级 = 2
-    基础 = 28837*1.1823*1.085 + 67288*1.1823*1.085 + 96125*1.1823*1.085
-    成长 = 8706*1.1823*1.085 + 20313 *1.1823*1.085 + 29019*1.1823*1.085
+    基础 = 1.065*(28837*1.1823*1.085 + 67288*1.1823*1.085 + 96125*1.1823*1.085)
+    成长 = 1.065*(8706*1.1823*1.085 + 20313 *1.1823*1.085 + 29019*1.1823*1.085)
     CD = 290
     演出时间 = 5.7
     关联技能 = ['无']
