+++ Characters/知源·缔造者.py
@@ -34,7 +34,7 @@ class 技能0(职业主动技能):
     等级上限 = 60
     基础等级 = 50
     data0 = [0, 11, 15, 20, 24, 29, 31, 35, 40, 44, 48, 53, 57, 61, 65, 69, 74, 78, 83, 85, 89, 94, 98, 102, 107, 110, 115, 119, 123, 128, 132, 137, 139, 143, 148, 152, 156, 160, 164, 169, 173, 177, 182, 186, 191, 193, 197, 202, 206, 210, 214, 218, 223, 227, 231, 236, 240, 242, 247, 251, 256, 260, 263, 268, 272, 277, 281, 285, 290, 294, 296]
-    data0=[(1+1.1244)*x*1.085 for x in data0]
+    data0=[(1+1.1244)*x*1.085*1.052 for x in data0]
 
     攻击次数 = 1
     能量 = 140
@@ -62,7 +62,7 @@ class 技能1(职业主动技能):
     等级上限 = 60
     基础等级 = 50
     data0 = [0, 109, 148, 187, 226, 263, 302, 341, 380, 418, 457, 494, 534, 573, 611, 650, 687, 727, 765, 804, 842, 882, 920, 958, 996, 1035, 1075, 1113, 1151, 1189, 1228, 1267, 1306, 1344, 1382, 1420, 1460, 1498, 1537, 1574, 1614, 1653, 1691, 1730, 1768, 1808, 1845, 1884, 1922, 1961, 2001, 2038, 2076, 2115, 2155, 2193, 2232, 2269, 2308, 2347, 2386, 2424, 2462, 2500, 2540, 2579, 2617, 2655, 2693, 2733, 2771]
-    data0=[1.213*1.085*x for x in data0]
+    data0=[1.213*1.085*x*1.052 for x in data0]
 
     攻击次数 = 1
     能量 = 140
@@ -85,7 +85,7 @@ class 技能2(职业主动技能):
     所在等级 = 10
     等级上限 = 60
     基础等级 = 48
-    data0 = [int(i*1.085) for i in [0, 78, 89, 100, 111, 124, 135, 146, 158, 169, 180, 193, 204, 215, 226, 238, 249, 260, 273, 284, 295, 307, 318, 329, 341, 353, 364, 375, 387, 398, 410, 422, 433, 444, 455, 467, 479, 490, 502, 513, 524, 536, 548, 559, 570, 582, 593, 604, 617, 628, 639, 651, 662, 673, 684, 697, 708, 719, 731, 742, 754, 766, 777, 788, 799, 811, 823, 834, 844, 856, 867]]
+    data0 = [int(i*1.085*1.05) for i in [0, 78, 89, 100, 111, 124, 135, 146, 158, 169, 180, 193, 204, 215, 226, 238, 249, 260, 273, 284, 295, 307, 318, 329, 341, 353, 364, 375, 387, 398, 410, 422, 433, 444, 455, 467, 479, 490, 502, 513, 524, 536, 548, 559, 570, 582, 593, 604, 617, 628, 639, 651, 662, 673, 684, 697, 708, 719, 731, 742, 754, 766, 777, 788, 799, 811, 823, 834, 844, 856, 867]]
 
 
     攻击次数 = 6
@@ -112,7 +112,7 @@ class 技能3(职业主动技能):
     基础等级 = 48
     # 冰盘攻击力：<data0>
     data0 = [0, 828, 950, 1071, 1191, 1313, 1433, 1555, 1675, 1797, 1917, 2038, 2160, 2281, 2401, 2523, 2643, 2764, 2885, 3007, 3127, 3247, 3368, 3489, 3611, 3731, 3853, 3973, 4094, 4214, 4337, 4457, 4578, 4699, 4821, 4941, 5063, 5183, 5304, 5424, 5547, 5667, 5788, 5909, 6030, 6150, 6273, 6393, 6514, 6634, 6754, 6877, 6997, 7118, 7239, 7360, 7480, 7603, 7723, 7844, 7964, 8086, 8207, 8328, 8449, 8570, 8690, 8813, 8933, 9054, 9174]
-    data0=[1.101*1.085*x for x in data0]
+    data0=[1.101*1.085*x*1.05 for x in data0]
 
     攻击次数 = 1
     # 能量消耗值：40
@@ -169,7 +169,7 @@ class 技能6(职业主动技能):
     基础等级 = 38
     # 烈焰飓风攻击力：<data0>
     data0 = [0, 457, 503, 550, 596, 641, 689, 735, 782, 827, 873, 921, 967, 1014, 1059, 1107, 1153, 1200, 1245, 1291, 1339, 1385, 1430, 1477, 1524, 1571, 1616, 1663, 1709, 1757, 1802, 1848, 1895, 1942, 1987, 2034, 2080, 2127, 2174, 2220, 2266, 2312, 2360, 2405, 2452, 2498, 2545, 2591, 2637, 2684, 2730, 2776, 2823, 2869, 2916, 2961, 3009, 3055, 3102, 3147, 3193, 3241, 3287, 3334, 3379, 3425, 3473, 3519, 3565, 3611, 3659]
-    data0=[1.153*1.085*x for x in data0]
+    data0=[1.153*1.085*x*1.051 for x in data0]
 
     攻击次数 = 12
     # 烈焰飓风持续时间：3秒
@@ -199,7 +199,7 @@ class 技能7(职业主动技能):
     等级上限 = 60
     基础等级 = 38
     data0 = [0, 2901, 3195, 3490, 3783, 4079, 4373, 4667, 4961, 5256, 5551, 5844, 6138, 6433, 6728, 7022, 7316, 7611, 7905, 8200, 8494, 8787, 9083, 9377, 9672, 9965, 10259, 10555, 10848, 11143, 11437, 11733, 12026, 12320, 12615, 12908, 13204, 13498, 13793, 14086, 14382, 14676, 14969, 15264, 15559, 15854, 16147, 16441, 16736, 17030, 17325, 17619, 17913, 18208, 18503, 18797, 19090, 19385, 19680, 19974, 20268, 20562, 20858, 21151, 21446, 21740, 22035, 22329, 22623, 22917, 23211]
-    data0=[1.19*1.085*x for x in data0]
+    data0=[1.19*1.085*x*1.052 for x in data0]
     攻击次数 = 1
 
     CD = 20.0
@@ -217,7 +217,7 @@ class 技能8(职业主动技能):
     等级上限 = 60
     基础等级 = 33
     data0 = [0, 375, 414, 450, 488, 527, 565, 603, 641, 680, 718, 757, 793, 832, 870, 908, 946, 985, 1023, 1060, 1099, 1137, 1174, 1212, 1251, 1289, 1327, 1365, 1404, 1442, 1480, 1517, 1555, 1593, 1631, 1670, 1708, 1746, 1784, 1823, 1861, 1898, 1936, 1975, 2013, 2050, 2089, 2127, 2166, 2203, 2241, 2279, 2317, 2355, 2394, 2432, 2470, 2508, 2547, 2585, 2621, 2660, 2698, 2736, 2774, 2813, 2851, 2890, 2927, 2965, 3003]
-    data0=[1.13*1.085 *x for x in data0]
+    data0=[1.13*1.085 *x*1.051 for x in data0]
 
     攻击次数 = 1
     CD = 25.0
@@ -249,7 +249,7 @@ class 技能9(职业主动技能):
     等级上限 = 60
     基础等级 = 33
     data0 = [0, 1575, 1735, 1895, 2055, 2215, 2375, 2536, 2695, 2856, 3015, 3176, 3335, 3494, 3654, 3814, 3974, 4134, 4294, 4454, 4615, 4774, 4935, 5094, 5253, 5413, 5573, 5733, 5893, 6053, 6213, 6373, 6533, 6694, 6853, 7014, 7172, 7332, 7492, 7652, 7812, 7972, 8132, 8292, 8452, 8612, 8773, 8931, 9090, 9251, 9411, 9571, 9731, 9891, 10051, 10211, 10371, 10531, 10691, 10850, 11010, 11171, 11330, 11490, 11650, 11810, 11970, 12130, 12290, 12450, 12609]
-    data0=[1.101*1.085*x for x in data0]
+    data0=[1.101*1.085*x*1.05 for x in data0]
 
     攻击次数 = 1
 
@@ -299,8 +299,8 @@ class 技能11(职业主动技能):
     data1 = [0, 10778, 12497, 14215, 15934, 17651, 19371, 21089, 22807, 24525, 26242, 27962, 29681, 31398, 33117, 34835, 36554, 38272, 39991, 41708, 43428, 45145, 46864, 48582, 50300, 52019, 53738, 55455, 57174, 58892, 60611, 62329, 64047, 65765, 67485, 69202, 70921, 72638, 74358, 76077, 77794, 79512, 81231, 82950, 84668, 86387, 88104, 89822, 91542, 93260, 94978, 96696, 98415, 100134, 101851, 103570, 105288, 107007, 108725, 110444, 112161, 113881, 115598, 117317, 119035, 120753, 122472, 124191, 125908, 127627, 129347]
 
 
-    data0=[1.3*1.085*x for x in data0]
-    data1=[1.3*1.085*x for x in data1]
+    data0=[1.3*1.085*x*1.101 for x in data0]
+    data1=[1.3*1.085*x*1.101 for x in data1]
 
     攻击次数 = 19
     攻击次数1 = 1
@@ -339,7 +339,7 @@ class 技能12(职业主动技能):
         ]
     ]
 
-    data0=[1.101*x for x in data0]
+    data0=[1.101*x*1.051 for x in data0]
 
     # 范围大小 1.2~1的倍率
     倍率 = 1.2
@@ -421,8 +421,8 @@ class 技能15(职业主动技能):
 
 
 
-    data0=[1.172*1.085*x for x in data0]
-    data1=[1.172*1.085*x for x in data1]
+    data0=[1.172*1.085*x*1.051 for x in data0]
+    data1=[1.172*1.085*x*1.051 for x in data1]
 
     攻击次数 = 30
     攻击次数1 = 1
@@ -450,7 +450,7 @@ class 技能16(职业主动技能):
     #    成长 = 25632.0
 
     data0 = [0, 19457, 23968, 28479, 32992, 37503, 42016, 46527, 51040, 55551, 60063, 64574, 69087, 73598, 78110, 82621, 87134, 91645, 96157, 100668, 105182, 109693, 114205, 118717, 123229, 127740, 132252, 136764, 141276, 145787, 150299, 154811, 159323, 163834, 168348, 172858, 177371, 181882, 186395, 190906, 195418, 199929, 204442, 208953, 213465, 217976, 222489, 227000, 231513, 236023, 240537, 245048, 249560, 254071, 258584, 263095, 267607, 272119, 276631, 281142, 285654, 290166, 294678, 299189, 303702, 308214, 312726, 317237, 321749, 326261, 330773]
-    data0=[1.15*1.085*x for x in data0]
+    data0=[1.15*1.085*x*1.052 for x in data0]
 
     攻击次数 = 6
     CD = 180.0
@@ -522,7 +522,7 @@ class 技能18(职业主动技能):
     #    成长 = 25632.0
     data0 = [0, 108124, 119092, 130060, 141030, 151998, 162969, 173937, 184907, 195875, 206846, 217814, 228782, 239752, 250720, 261691, 272659, 283630, 294597, 305568, 316536, 327507, 338475, 349442, 360413, 371381, 382352, 393320, 404290, 415258, 426229, 437197, 448164, 459135, 470103, 481074, 492042, 503012, 513980, 524951, 535919, 546887, 557857, 568825, 579796, 590764, 601735, 612702, 623673, 634641, 645609, 656580, 667547, 678518, 689486, 700457, 711424, 722395, 733363, 744331, 755302, 766269, 777240, 788208, 799179, 810147, 821117, 832085, 843053, 854024, 864992]
 
-    data0=[1.191*1.085*x for x in data0]
+    data0=[1.191*1.085*x*1.051 for x in data0]
 
     攻击次数 = 1
     CD = 60.0
