+++ Characters/知源·血法师.py
@@ -129,7 +129,7 @@ class 技能7(职业主动技能):
     基础等级 =43
     #基础 = 2993.1012727272
     #成长 = 337.716727272
-    data0 = [ int(i*1.193*1.085) for i in [0, 1666, 1835, 2004, 2173, 2342, 2511, 2681, 2850, 3019, 3188, 3357, 3526, 3695, 3864, 4033, 4202, 4371, 4540, 4709, 4879, 5048, 5217, 5386, 5555, 5724, 5893, 6062, 6231, 6400, 6569, 6738, 6907, 7077, 7246, 7415, 7584, 7753, 7922, 8091, 8260, 8429, 8598, 8767, 8936, 9105, 9275, 9444, 9613, 9782, 9951, 10120, 10289, 10458, 10627, 10796, 10965, 11134, 11303, 11473, 11642, 11811, 11980, 12149, 12318, 12487, 12656, 12825, 12994, 13163, 13332]]
+    data0 = [ int(i*1.193*1.085*1.051) for i in [0, 1666, 1835, 2004, 2173, 2342, 2511, 2681, 2850, 3019, 3188, 3357, 3526, 3695, 3864, 4033, 4202, 4371, 4540, 4709, 4879, 5048, 5217, 5386, 5555, 5724, 5893, 6062, 6231, 6400, 6569, 6738, 6907, 7077, 7246, 7415, 7584, 7753, 7922, 8091, 8260, 8429, 8598, 8767, 8936, 9105, 9275, 9444, 9613, 9782, 9951, 10120, 10289, 10458, 10627, 10796, 10965, 11134, 11303, 11473, 11642, 11811, 11980, 12149, 12318, 12487, 12656, 12825, 12994, 13163, 13332]]
     攻击次数 = 2
     CD = 6
     TP成长 = 0.10
@@ -142,7 +142,7 @@ class 技能8(职业主动技能):
     基础等级 =41
     #基础 = 3293.18181818
     #成长 = 372.58741259
-    data0 = [ int(i*1.3*1.085) for i in [0, 122, 135, 147, 159, 172, 185, 196, 209, 222, 234, 246, 259, 271, 284, 296, 308, 321, 334, 345, 358, 371, 383, 395, 408, 420, 433, 445, 457, 470, 483, 495, 507, 520, 532, 545, 557, 569, 582, 595, 606, 619, 632, 644, 656, 669, 681, 694, 706, 718, 731, 744, 755, 768, 781, 793, 805, 818, 830, 843, 855, 867, 880, 893, 905, 917, 930, 942, 955, 967, 979]]
+    data0 = [ int(i*1.3*1.085*1.053) for i in [0, 122, 135, 147, 159, 172, 185, 196, 209, 222, 234, 246, 259, 271, 284, 296, 308, 321, 334, 345, 358, 371, 383, 395, 408, 420, 433, 445, 457, 470, 483, 495, 507, 520, 532, 545, 557, 569, 582, 595, 606, 619, 632, 644, 656, 669, 681, 694, 706, 718, 731, 744, 755, 768, 781, 793, 805, 818, 830, 843, 855, 867, 880, 893, 905, 917, 930, 942, 955, 967, 979]]
     攻击次数 = 30
     CD = 6
     TP成长 = 0.10
@@ -155,7 +155,7 @@ class 技能9(职业主动技能):
     基础等级 =41
     #基础 = 4179
     #成长 = 472.5
-    data0 = [ int(i*1.114*1.085) for i in [0, 931, 1025, 1120, 1214, 1309, 1403, 1498, 1592, 1687, 1781, 1876, 1970, 2065, 2159, 2254, 2348, 2443, 2537, 2632, 2726, 2821, 2915, 3010, 3104, 3199, 3293, 3388, 3482, 3577, 3671, 3766, 3860, 3955, 4049, 4144, 4238, 4333, 4427, 4522, 4616, 4711, 4805, 4900, 4994, 5089, 5183, 5278, 5372, 5467, 5561, 5656, 5750, 5845, 5939, 6034, 6128, 6223, 6317, 6412, 6506, 6601, 6695, 6790, 6884, 6979, 7073, 7168, 7262, 7357, 7451]]
+    data0 = [ int(i*1.114*1.085*1.051) for i in [0, 931, 1025, 1120, 1214, 1309, 1403, 1498, 1592, 1687, 1781, 1876, 1970, 2065, 2159, 2254, 2348, 2443, 2537, 2632, 2726, 2821, 2915, 3010, 3104, 3199, 3293, 3388, 3482, 3577, 3671, 3766, 3860, 3955, 4049, 4144, 4238, 4333, 4427, 4522, 4616, 4711, 4805, 4900, 4994, 5089, 5183, 5278, 5372, 5467, 5561, 5656, 5750, 5845, 5939, 6034, 6128, 6223, 6317, 6412, 6506, 6601, 6695, 6790, 6884, 6979, 7073, 7168, 7262, 7357, 7451]]
     攻击次数 = 5
     CD = 8
     TP成长 = 0.10
@@ -201,7 +201,7 @@ class 技能12(职业主动技能):
     基础等级 =36
     #基础 = 9392.703939
     #成长 = 1061.7061888
-    data0 = [ int(i*1.161*1.085) for i in [0, 1046, 1152, 1259, 1365, 1471, 1577, 1683, 1790, 1896, 2002, 2108, 2214, 2320, 2427, 2533, 2639, 2745, 2851, 2958, 3064, 3170, 3276, 3382, 3489, 3595, 3701, 3807, 3913, 4020, 4126, 4232, 4338, 4444, 4550, 4657, 4763, 4869, 4975, 5081, 5188, 5294, 5400, 5506, 5612, 5719, 5825, 5931, 6037, 6143, 6249, 6356, 6462, 6568, 6674, 6780, 6887, 6993, 7099, 7205, 7311, 7418, 7524, 7630, 7736, 7842, 7949, 8055, 8161, 8267, 8373]]
+    data0 = [ int(i*1.161*1.085*1.053) for i in [0, 1046, 1152, 1259, 1365, 1471, 1577, 1683, 1790, 1896, 2002, 2108, 2214, 2320, 2427, 2533, 2639, 2745, 2851, 2958, 3064, 3170, 3276, 3382, 3489, 3595, 3701, 3807, 3913, 4020, 4126, 4232, 4338, 4444, 4550, 4657, 4763, 4869, 4975, 5081, 5188, 5294, 5400, 5506, 5612, 5719, 5825, 5931, 6037, 6143, 6249, 6356, 6462, 6568, 6674, 6780, 6887, 6993, 7099, 7205, 7311, 7418, 7524, 7630, 7736, 7842, 7949, 8055, 8161, 8267, 8373]]
     攻击次数 = 10
     CD =18
     TP成长 = 0.10
@@ -228,7 +228,7 @@ class 技能13(职业主动技能):
     #数组不正确，但仅体现在游戏数据中，计算器最终得出结果无异常
     #基础 = 12175.36181818/31*30
     #成长 = 1333.46433566/31*30
-    data0 = [ int(i*1.193*1.085) for i in [0, 437, 481, 526, 570, 615, 659, 703, 748, 792, 837, 881, 925, 970, 1014, 1059, 1103, 1147, 1192, 1236, 1281, 1325, 1369, 1414, 1458, 1502, 1547, 1591, 1636, 1680, 1724, 1769, 1813, 1858, 1902, 1946, 1991, 2035, 2080, 2124, 2168, 2213, 2257, 2302, 2346, 2390, 2435, 2479, 2523, 2568, 2612, 2657, 2701, 2745, 2790, 2834, 2879, 2923, 2967, 3012, 3056, 3101, 3145, 3189, 3234, 3278, 3323, 3367, 3411, 3456, 3500]]
+    data0 = [ int(i*1.193*1.085*1.051) for i in [0, 437, 481, 526, 570, 615, 659, 703, 748, 792, 837, 881, 925, 970, 1014, 1059, 1103, 1147, 1192, 1236, 1281, 1325, 1369, 1414, 1458, 1502, 1547, 1591, 1636, 1680, 1724, 1769, 1813, 1858, 1902, 1946, 1991, 2035, 2080, 2124, 2168, 2213, 2257, 2302, 2346, 2390, 2435, 2479, 2523, 2568, 2612, 2657, 2701, 2745, 2790, 2834, 2879, 2923, 2967, 3012, 3056, 3101, 3145, 3189, 3234, 3278, 3323, 3367, 3411, 3456, 3500]]
     攻击次数 = 30
     CD =20
     TP成长 = 0.10
@@ -255,7 +255,7 @@ class 技能14(职业主动技能):
     #基础 = 23208.8198030
     #成长 = 2620.41055594
     #吸血物理攻击力：<data0>%
-    data0 = [ int(i*1.192*1.085) for i in [0, 25837, 28458, 31079, 33701, 36322, 38943, 41564, 44185, 46807, 49428, 52049, 54670, 57291, 59913, 62534, 65155, 67776, 70397, 73018, 75640, 78261, 80882, 83503, 86124, 88746, 91367, 93988, 96609, 99230, 101852, 104473, 107094, 109715, 112336, 114958, 117579, 120200, 122821, 125442, 128064, 130685, 133306, 135927, 138548, 141170, 143791, 146412, 149033, 151654, 154275, 156897, 159518, 162139, 164760, 167381, 170003, 172624, 175245, 177866, 180487, 183109, 185730, 188351, 190972, 193593, 196215, 198836, 201457, 204078, 206699]]
+    data0 = [ int(i*1.192*1.085*1.052) for i in [0, 25837, 28458, 31079, 33701, 36322, 38943, 41564, 44185, 46807, 49428, 52049, 54670, 57291, 59913, 62534, 65155, 67776, 70397, 73018, 75640, 78261, 80882, 83503, 86124, 88746, 91367, 93988, 96609, 99230, 101852, 104473, 107094, 109715, 112336, 114958, 117579, 120200, 122821, 125442, 128064, 130685, 133306, 135927, 138548, 141170, 143791, 146412, 149033, 151654, 154275, 156897, 159518, 162139, 164760, 167381, 170003, 172624, 175245, 177866, 180487, 183109, 185730, 188351, 190972, 193593, 196215, 198836, 201457, 204078, 206699]]
     CD =40
     TP成长 = 0.10
     TP上限 = 5
@@ -322,9 +322,9 @@ class 技能17(职业主动技能):
     #基础 = 25260.95
     #成长 = 2852.003846153
     #蝙蝠群冲击波攻击力：<data0>%
-    data0 = [ int(i*1.265*1.085) for i in [0, 11245, 12386, 13527, 14667, 15808, 16949, 18090, 19231, 20372, 21512, 22653, 23794, 24935, 26076, 27217, 28358, 29498, 30639, 31780, 32921, 34062, 35203, 36343, 37484, 38625, 39766, 40907, 42048, 43188, 44329, 45470, 46611, 47752, 48893, 50034, 51174, 52315, 53456, 54597, 55738]]
+    data0 = [ int(i*1.265*1.085*1.053) for i in [0, 11245, 12386, 13527, 14667, 15808, 16949, 18090, 19231, 20372, 21512, 22653, 23794, 24935, 26076, 27217, 28358, 29498, 30639, 31780, 32921, 34062, 35203, 36343, 37484, 38625, 39766, 40907, 42048, 43188, 44329, 45470, 46611, 47752, 48893, 50034, 51174, 52315, 53456, 54597, 55738]]
     #血气之翼物理攻击力：<data1>%
-    data1 = [ int(i*1.265*1.085) for i in [0, 16869, 18579, 20290, 22001, 23713, 25424, 27135, 28846, 30558, 32269, 33980, 35691, 37403, 39114, 40825, 42537, 44248, 45959, 47670, 49382, 51093, 52804, 54515, 56227, 57938, 59649, 61360, 63072, 64783, 66494, 68206, 69917, 71628, 73339, 75050, 76762, 78473, 80185, 81895, 83607]]
+    data1 = [ int(i*1.265*1.085*1.053) for i in [0, 16869, 18579, 20290, 22001, 23713, 25424, 27135, 28846, 30558, 32269, 33980, 35691, 37403, 39114, 40825, 42537, 44248, 45959, 47670, 49382, 51093, 52804, 54515, 56227, 57938, 59649, 61360, 63072, 64783, 66494, 68206, 69917, 71628, 73339, 75050, 76762, 78473, 80185, 81895, 83607]]
     攻击次数2 = 1
     CD =50
     TP成长 = 0.10
@@ -349,10 +349,10 @@ class 技能18(职业主动技能):
     #基础 = 30218.85
     #成长 = 5734.946153849
     #撕咬攻击力：<data0>%
-    data0 = [ int(i*1.227*1.085) for i in [0, 899, 1042, 1186, 1329, 1471, 1616, 1759, 1902, 2046, 2189, 2332, 2476, 2619, 2761, 2906, 3049, 3192, 3336, 3478, 3623, 3766, 3909, 4054, 4197, 4340, 4483, 4626, 4769, 4914, 5056, 5199, 5343, 5486, 5629, 5773, 5916, 6059, 6204, 6346, 6489, 6633, 6776, 6920, 7063, 7206, 7350, 7493, 7636, 7780, 7923, 8066, 8210, 8353, 8496, 8640, 8783, 8926, 9070, 9213, 9356, 9500, 9643, 9787, 9930, 10073, 10217, 10360, 10503, 10647, 10790]]
+    data0 = [ int(i*1.227*1.085*1.05) for i in [0, 899, 1042, 1186, 1329, 1471, 1616, 1759, 1902, 2046, 2189, 2332, 2476, 2619, 2761, 2906, 3049, 3192, 3336, 3478, 3623, 3766, 3909, 4054, 4197, 4340, 4483, 4626, 4769, 4914, 5056, 5199, 5343, 5486, 5629, 5773, 5916, 6059, 6204, 6346, 6489, 6633, 6776, 6920, 7063, 7206, 7350, 7493, 7636, 7780, 7923, 8066, 8210, 8353, 8496, 8640, 8783, 8926, 9070, 9213, 9356, 9500, 9643, 9787, 9930, 10073, 10217, 10360, 10503, 10647, 10790]]
     攻击次数 = 20
     #冲向地面攻击力：<data1>%
-    data1 = [ int(i*1.227*1.085) for i in [0, 17984, 20851, 23718, 26586, 29453, 32320, 35188, 38055, 40922, 43789, 46655, 49522, 52390, 55257, 58124, 60992, 63859, 66726, 69594, 72461, 75327, 78194, 81061, 83928, 86796, 89663, 92530, 95398, 98265, 101132, 104000, 106866, 109732, 112600, 115467, 118334, 121202, 124069, 126936, 129804, 132670, 135538, 138405, 141272, 144139, 147006, 149873, 152740, 155608, 158475, 161342, 164209, 167076, 169943, 172811, 175678, 178545, 181412, 184279, 187146, 190013, 192881, 195748, 198615, 201482, 204349, 207216, 210084, 212951, 215818]]
+    data1 = [ int(i*1.227*1.085*1.05) for i in [0, 17984, 20851, 23718, 26586, 29453, 32320, 35188, 38055, 40922, 43789, 46655, 49522, 52390, 55257, 58124, 60992, 63859, 66726, 69594, 72461, 75327, 78194, 81061, 83928, 86796, 89663, 92530, 95398, 98265, 101132, 104000, 106866, 109732, 112600, 115467, 118334, 121202, 124069, 126936, 129804, 132670, 135538, 138405, 141272, 144139, 147006, 149873, 152740, 155608, 158475, 161342, 164209, 167076, 169943, 172811, 175678, 178545, 181412, 184279, 187146, 190013, 192881, 195748, 198615, 201482, 204349, 207216, 210084, 212951, 215818]]
     攻击次数2 = 1
     CD =30
     是否有护石 = 1
@@ -371,7 +371,7 @@ class 技能19(职业主动技能):
     基础等级 =13
     #基础 = 60756.9115151488
     #成长 = 6859.7572027936
-    data0 = [ int(i*1.225*1.085) for i in [0, 67617, 74477, 81337, 88196, 95056, 101916, 108776, 115634, 122494, 129354, 136214, 143074, 149933, 156793, 163653, 170513, 177372, 184232, 191092, 197952, 204811, 211671, 218530, 225391, 232250, 239110, 245969, 252830, 259689, 266549, 273409, 280269, 287129, 293988, 300848, 307707, 314568, 321427, 328287, 335145]]
+    data0 = [ int(i*1.225*1.085*1.053) for i in [0, 67617, 74477, 81337, 88196, 95056, 101916, 108776, 115634, 122494, 129354, 136214, 143074, 149933, 156793, 163653, 170513, 177372, 184232, 191092, 197952, 204811, 211671, 218530, 225391, 232250, 239110, 245969, 252830, 259689, 266549, 273409, 280269, 287129, 293988, 300848, 307707, 314568, 321427, 328287, 335145]]
     CD =50
     是否有护石 = 1
     护石选项 = ['圣痕']
@@ -401,7 +401,7 @@ class 技能21(职业主动技能):
     所在等级 = 95
     等级上限 = 40
     基础等级 = 6
-    data0 = [ int(i*1.186*1.085) for i in [0, 116057, 127831, 139604, 151378, 163152, 174926, 186700, 198474, 210248, 222022, 233796, 245570, 257344, 269117, 280891, 292665, 304439, 316213, 327987, 339761, 351535, 363309, 375083, 386857, 398630, 410404, 422178, 433952, 445726, 457500, 469274, 481048, 492822, 504596, 516370, 528144, 539917, 551691, 563465, 575239, 587013, 598787, 610561, 622335, 634109, 645883, 657657, 669430, 681204, 692978, 704752, 716526, 728300, 740074, 751848, 763622, 775396, 787170, 798943, 810717, 822491, 834265, 846039, 857813, 869587, 881361, 893135, 904909, 916683, 928456]]
+    data0 = [ int(i*1.186*1.085*1.052) for i in [0, 116057, 127831, 139604, 151378, 163152, 174926, 186700, 198474, 210248, 222022, 233796, 245570, 257344, 269117, 280891, 292665, 304439, 316213, 327987, 339761, 351535, 363309, 375083, 386857, 398630, 410404, 422178, 433952, 445726, 457500, 469274, 481048, 492822, 504596, 516370, 528144, 539917, 551691, 563465, 575239, 587013, 598787, 610561, 622335, 634109, 645883, 657657, 669430, 681204, 692978, 704752, 716526, 728300, 740074, 751848, 763622, 775396, 787170, 798943, 810717, 822491, 834265, 846039, 857813, 869587, 881361, 893135, 904909, 916683, 928456]]
     CD = 60.0
 
 class 技能22(职业主动技能):
