+++ Characters/神启·异端审判者.py
@@ -220,13 +220,13 @@ class 技能10(职业主动技能):
 
     #基础 = 1209  - 123
     #成长 = 123
-    data0 = [ int(i*1.1389*1.085) for i in [0, 1209, 1332, 1455, 1578, 1700, 1823, 1946, 2069, 2191, 2314, 2437, 2560, 2682, 2805, 2928, 3051, 3173, 3296, 3419, 3542, 3664, 3787, 3910, 4033, 4155, 4278, 4401, 4524, 4647, 4769, 4892, 5015, 5138, 5260, 5383, 5506, 5629, 5751, 5874, 5997, 6120, 6242, 6365, 6488, 6611, 6733, 6856, 6979, 7102, 7224, 7347, 7470, 7593, 7715, 7838, 7961, 8084, 8206, 8329, 8452, 8575, 8697, 8820, 8943, 9066, 9188, 9311, 9434, 9557, 9679]]
+    data0 = [ int(i*1.1389*1.085*1.051) for i in [0, 1209, 1332, 1455, 1578, 1700, 1823, 1946, 2069, 2191, 2314, 2437, 2560, 2682, 2805, 2928, 3051, 3173, 3296, 3419, 3542, 3664, 3787, 3910, 4033, 4155, 4278, 4401, 4524, 4647, 4769, 4892, 5015, 5138, 5260, 5383, 5506, 5629, 5751, 5874, 5997, 6120, 6242, 6365, 6488, 6611, 6733, 6856, 6979, 7102, 7224, 7347, 7470, 7593, 7715, 7838, 7961, 8084, 8206, 8329, 8452, 8575, 8697, 8820, 8943, 9066, 9188, 9311, 9434, 9557, 9679]]
     攻击次数 = 1
 
     #基础2 = 10903 -1106
     #成长2 = 1106
     攻击次数2 = 1
-    data1 = [ int(i*1.137*1.085) for i in [0, 10903, 12009, 13116, 14222, 15328, 16434, 17540, 18646, 19753, 20859, 21965, 23071, 24177, 25284, 26390, 27496, 28602, 29708, 30814, 31921, 33027, 34133, 35239, 36345, 37451, 38558, 39664, 40770, 41876, 42982, 44089, 45195, 46301, 47407, 48513, 49619, 50726, 51832, 52938, 54044, 55150, 56256, 57363, 58469, 59575, 60681, 61787, 62893, 64000, 65106, 66212, 67318, 68424, 69531, 70637, 71743, 72849, 73955, 75061, 76168, 77274, 78380, 79486, 80592, 81698, 82805, 83911, 85017, 86123, 87229]]
+    data1 = [ int(i*1.137*1.085*1.051) for i in [0, 10903, 12009, 13116, 14222, 15328, 16434, 17540, 18646, 19753, 20859, 21965, 23071, 24177, 25284, 26390, 27496, 28602, 29708, 30814, 31921, 33027, 34133, 35239, 36345, 37451, 38558, 39664, 40770, 41876, 42982, 44089, 45195, 46301, 47407, 48513, 49619, 50726, 51832, 52938, 54044, 55150, 56256, 57363, 58469, 59575, 60681, 61787, 62893, 64000, 65106, 66212, 67318, 68424, 69531, 70637, 71743, 72849, 73955, 75061, 76168, 77274, 78380, 79486, 80592, 81698, 82805, 83911, 85017, 86123, 87229]]
     # 默认武器处于点火状态，怪物处于焚烧状态，否则data1 = [0, 10889, 11994, 13099, 14204, 15308, 16413, 17518, 18623, 19727, 20832, 21937, 23042, 24146, 25251, 26356, 27461, 28565, 29670, 30775, 31880, 32984, 34089, 35194, 36299, 37403, 38508, 39613, 40718, 41823, 42927, 44032, 45137, 46242, 47346, 48451, 49556, 50661, 51765, 52870, 53975, 55080, 56184, 57289, 58394, 59499, 60603, 61708, 62813, 63918, 65022, 66127, 67232, 68337, 69441, 70546, 71651, 72756, 73860, 74965, 76070, 77175, 78280, 79384, 80489, 81594, 82699, 83803, 84908, 86013, 87118]
     CD = 20
     TP成长 = 0.10
@@ -252,7 +252,7 @@ class 技能11(职业主动技能):
     等级上限 = 60
     基础等级 = 36
     #基础 = 9139 - 928
-    data0 = [ int(i*1.203*1.085) for i in [0,9139, 10067, 10996, 11922, 12850, 13777, 14705, 15631, 16559, 17486, 18412, 19340, 20269, 21196, 22124, 23050, 23978, 24905, 25831, 26759, 27687, 28615, 29543, 30469, 31397, 32324, 33251, 34178, 35106, 36032, 36961, 37888, 38816, 39743, 40670, 41597, 42525, 43451, 44379, 45307, 46235, 47162, 48089, 49016, 49944, 50870, 51798, 52725, 53654, 54581, 55508, 56435, 57363, 58289, 59217, 60144, 61072, 61998, 62927, 63854, 64782, 65708, 66636, 67563, 68491, 69417, 70345, 71273, 72201, 73127]]
+    data0 = [ int(i*1.203*1.085*1.053) for i in [0,9139, 10067, 10996, 11922, 12850, 13777, 14705, 15631, 16559, 17486, 18412, 19340, 20269, 21196, 22124, 23050, 23978, 24905, 25831, 26759, 27687, 28615, 29543, 30469, 31397, 32324, 33251, 34178, 35106, 36032, 36961, 37888, 38816, 39743, 40670, 41597, 42525, 43451, 44379, 45307, 46235, 47162, 48089, 49016, 49944, 50870, 51798, 52725, 53654, 54581, 55508, 56435, 57363, 58289, 59217, 60144, 61072, 61998, 62927, 63854, 64782, 65708, 66636, 67563, 68491, 69417, 70345, 71273, 72201, 73127]]
     #2020年11月29日微调
     #成长 = 928
     CD = 15
@@ -272,10 +272,10 @@ class 技能12(职业主动技能):
     TP上限 = 5
     是否有护石 = 1
 
-    data0 = [ int(i*1.2027*1.085) for i in [0,1687, 1859, 2030, 2202, 2372, 2544, 2715, 2887, 3059, 3230, 3401, 3571, 3743, 3915, 4086, 4258, 4428, 4600, 4772, 4943, 5115, 5284, 5456, 5627, 5799, 5971, 6142, 6313, 6484, 6656, 6828, 6999, 7171, 7340, 7512, 7684, 7855, 8027, 8198, 8369, 8540, 8712, 8884, 9055, 9227, 9396, 9568, 9740, 9911, 10083, 10254, 10425, 10596, 10768, 10940, 11110, 11281, 11452, 11624, 11796, 11967, 12139, 12309, 12481, 12653, 12824, 12995, 13166, 13337, 13508]]
+    data0 = [ int(i*1.2027*1.085*1.053) for i in [0,1687, 1859, 2030, 2202, 2372, 2544, 2715, 2887, 3059, 3230, 3401, 3571, 3743, 3915, 4086, 4258, 4428, 4600, 4772, 4943, 5115, 5284, 5456, 5627, 5799, 5971, 6142, 6313, 6484, 6656, 6828, 6999, 7171, 7340, 7512, 7684, 7855, 8027, 8198, 8369, 8540, 8712, 8884, 9055, 9227, 9396, 9568, 9740, 9911, 10083, 10254, 10425, 10596, 10768, 10940, 11110, 11281, 11452, 11624, 11796, 11967, 12139, 12309, 12481, 12653, 12824, 12995, 13166, 13337, 13508]]
     攻击次数 = 5
 
-    data1 = [ int(i*1.20295*1.085) for i in [0,5627, 6200, 6770, 7340, 7912, 8483, 9055, 9625, 10197, 10768, 11338, 11910, 12481, 13052, 13623, 14195, 14765, 15336, 15908, 16479, 17050, 17621, 18193, 18763, 19333, 19906, 20475, 21047, 21618, 22189, 22760, 23331, 23903, 24473, 25045, 25616, 26187, 26758, 27329, 27900, 28471, 29043, 29614, 30185, 30756, 31327, 31898, 32469, 33041, 33611, 34183, 34753, 35323, 35895, 36466, 37038, 37608, 38180, 38751, 39321, 39893, 40464, 41035, 41606, 42178, 42748, 43319, 43891, 44462, 45033]]
+    data1 = [ int(i*1.20295*1.085*1.053) for i in [0,5627, 6200, 6770, 7340, 7912, 8483, 9055, 9625, 10197, 10768, 11338, 11910, 12481, 13052, 13623, 14195, 14765, 15336, 15908, 16479, 17050, 17621, 18193, 18763, 19333, 19906, 20475, 21047, 21618, 22189, 22760, 23331, 23903, 24473, 25045, 25616, 26187, 26758, 27329, 27900, 28471, 29043, 29614, 30185, 30756, 31327, 31898, 32469, 33041, 33611, 34183, 34753, 35323, 35895, 36466, 37038, 37608, 38180, 38751, 39321, 39893, 40464, 41035, 41606, 42178, 42748, 43319, 43891, 44462, 45033]]
     攻击次数2 = 1
 
     护石选项 = ['魔界', '圣痕']
@@ -297,7 +297,7 @@ class 技能13(职业主动技能):
     基础等级 = 31
     #基础 = 23205 -2354.5
     #成长 = 2354.5
-    data0 = [ int(i*1.14*1.085) for i in [0, 23205, 25560, 27914, 30268, 32622, 34977, 37331, 39685, 42039, 44394, 46748, 49102, 51456, 53810, 56165, 58519, 60873, 63227, 65582, 67936, 70290, 72644, 74998, 77353, 79707, 82061, 84415, 86770, 89124, 91478, 93832, 96187, 98541, 100895, 103249, 105603, 107958, 110312, 112666, 115020, 117375, 119729, 122083, 124437, 126791, 129146, 131500, 133854, 136208, 138563, 140917, 143271, 145625, 147980, 150334, 152688, 155042, 157396, 159751, 162105, 164459, 166813, 169168, 171522, 173876, 176230, 178584, 180939, 183293, 185647]]
+    data0 = [ int(i*1.14*1.085*1.052) for i in [0, 23205, 25560, 27914, 30268, 32622, 34977, 37331, 39685, 42039, 44394, 46748, 49102, 51456, 53810, 56165, 58519, 60873, 63227, 65582, 67936, 70290, 72644, 74998, 77353, 79707, 82061, 84415, 86770, 89124, 91478, 93832, 96187, 98541, 100895, 103249, 105603, 107958, 110312, 112666, 115020, 117375, 119729, 122083, 124437, 126791, 129146, 131500, 133854, 136208, 138563, 140917, 143271, 145625, 147980, 150334, 152688, 155042, 157396, 159751, 162105, 164459, 166813, 169168, 171522, 173876, 176230, 178584, 180939, 183293, 185647]]
     CD = 45
     TP成长 = 0.10
     TP上限 = 5
@@ -353,10 +353,10 @@ class 技能15(职业主动技能):
     TP上限 = 5
     是否有护石 = 1
 
-    data0 =[ int(i*1.203*1.085) for i in [0,1639, 1805, 1972, 2139, 2304, 2471, 2637, 2803, 2970, 3136, 3302, 3470, 3635, 3803, 3968, 4134, 4301, 4468, 4634, 4801, 4967, 5132, 5299, 5465, 5633, 5799, 5965, 6132, 6298, 6463, 6630, 6796, 6963, 7130, 7296, 7463, 7629, 7794, 7961, 8128, 8294, 8461, 8627, 8793, 8960, 9125, 9294, 9459, 9625, 9792, 9958, 10124, 10292, 10458, 10623, 10790, 10956, 11122, 11289, 11455, 11623, 11789, 11954, 12121, 12287, 12453, 12620, 12787, 12953, 13120]]
+    data0 =[ int(i*1.203*1.085*1.053) for i in [0,1639, 1805, 1972, 2139, 2304, 2471, 2637, 2803, 2970, 3136, 3302, 3470, 3635, 3803, 3968, 4134, 4301, 4468, 4634, 4801, 4967, 5132, 5299, 5465, 5633, 5799, 5965, 6132, 6298, 6463, 6630, 6796, 6963, 7130, 7296, 7463, 7629, 7794, 7961, 8128, 8294, 8461, 8627, 8793, 8960, 9125, 9294, 9459, 9625, 9792, 9958, 10124, 10292, 10458, 10623, 10790, 10956, 11122, 11289, 11455, 11623, 11789, 11954, 12121, 12287, 12453, 12620, 12787, 12953, 13120]]
     攻击次数 = 9
 
-    data1 = [ int(i*1.203*1.085) for i in [0,3461, 3813, 4164, 4515, 4866, 5218, 5569, 5920, 6271, 6622, 6974, 7325, 7677, 8027, 8378, 8730, 9081, 9433, 9783, 10136, 10486, 10837, 11189, 11539, 11892, 12242, 12594, 12945, 13295, 13648, 13998, 14351, 14701, 15053, 15404, 15756, 16107, 16457, 16809, 17160, 17512, 17863, 18215, 18565, 18916, 19268, 19619, 19971, 20321, 20673, 21024, 21375, 21727, 22077, 22429, 22780, 23132, 23483, 23833, 24185, 24536, 24888, 25239, 25591, 25941, 26293, 26644, 26995, 27347, 27697]]
+    data1 = [ int(i*1.203*1.085*1.053) for i in [0,3461, 3813, 4164, 4515, 4866, 5218, 5569, 5920, 6271, 6622, 6974, 7325, 7677, 8027, 8378, 8730, 9081, 9433, 9783, 10136, 10486, 10837, 11189, 11539, 11892, 12242, 12594, 12945, 13295, 13648, 13998, 14351, 14701, 15053, 15404, 15756, 16107, 16457, 16809, 17160, 17512, 17863, 18215, 18565, 18916, 19268, 19619, 19971, 20321, 20673, 21024, 21375, 21727, 22077, 22429, 22780, 23132, 23483, 23833, 24185, 24536, 24888, 25239, 25591, 25941, 26293, 26644, 26995, 27347, 27697]]
     攻击次数2 = 1
 
     护石选项 = ['魔界', '圣痕']
@@ -380,7 +380,7 @@ class 技能16(职业主动技能):
 
     #基础 = 28032 - 2844
     #成长 = 2844
-    data0 = [ int(i*1.1953*1.085) for i in [0, 28032, 30875, 33719, 36563, 39407, 42251, 45095, 47938, 50782, 53626, 56470, 59314, 62158, 65001, 67845, 70689, 73533, 76377, 79221, 82064, 84908, 87752, 90596, 93440, 96284, 99127, 101971, 104815, 107659, 110503, 113347, 116190, 119034, 121878, 124722, 127566, 130410, 133254, 136097, 138941, 141785, 144629, 147473, 150317, 153160, 156004, 158848, 161692, 164536, 167380, 170223, 173067, 175911, 178755, 181599, 184443, 187286, 190130, 192974, 195818, 198662, 201506, 204349, 207193, 210037, 212881, 215725, 218569, 221412, 224256]]
+    data0 = [ int(i*1.1953*1.085*1.05) for i in [0, 28032, 30875, 33719, 36563, 39407, 42251, 45095, 47938, 50782, 53626, 56470, 59314, 62158, 65001, 67845, 70689, 73533, 76377, 79221, 82064, 84908, 87752, 90596, 93440, 96284, 99127, 101971, 104815, 107659, 110503, 113347, 116190, 119034, 121878, 124722, 127566, 130410, 133254, 136097, 138941, 141785, 144629, 147473, 150317, 153160, 156004, 158848, 161692, 164536, 167380, 170223, 173067, 175911, 178755, 181599, 184443, 187286, 190130, 192974, 195818, 198662, 201506, 204349, 207193, 210037, 212881, 215725, 218569, 221412, 224256]]
     CD = 40
     TP成长 = 0.10
     TP上限 = 5
@@ -400,9 +400,9 @@ class 技能17(职业主动技能):
     等级上限 = 40
     基础等级 = 16
 
-    data0 = [ int(i*1.1959*1.085) for i in [0,2873, 3165, 3456, 3749, 4040, 4332, 4623, 4916, 5207, 5499, 5790, 6081, 6373, 6665, 6956, 7248, 7540, 7832, 8123, 8415, 8707, 8999, 9290, 9582, 9874, 10164, 10456, 10747, 11039, 11331, 11623, 11914, 12206, 12498, 12790, 13081, 13373, 13665, 13957, 14247, 14539, 14830, 15123, 15414, 15706, 15997, 16290, 16581, 16873, 17164, 17457, 17748, 18040, 18330, 18621, 18914, 19205, 19497, 19788, 20081, 20372, 20664, 20955, 21248, 21539, 21831, 22122, 22413, 22705, 22996]]
+    data0 = [ int(i*1.1959*1.085*1.051) for i in [0,2873, 3165, 3456, 3749, 4040, 4332, 4623, 4916, 5207, 5499, 5790, 6081, 6373, 6665, 6956, 7248, 7540, 7832, 8123, 8415, 8707, 8999, 9290, 9582, 9874, 10164, 10456, 10747, 11039, 11331, 11623, 11914, 12206, 12498, 12790, 13081, 13373, 13665, 13957, 14247, 14539, 14830, 15123, 15414, 15706, 15997, 16290, 16581, 16873, 17164, 17457, 17748, 18040, 18330, 18621, 18914, 19205, 19497, 19788, 20081, 20372, 20664, 20955, 21248, 21539, 21831, 22122, 22413, 22705, 22996]]
     攻击次数 = 3
-    data1 = [ int(i*1.1959*1.085) for i in [0,45777, 50423, 55067, 59712, 64356, 68999, 73643, 78289, 82933, 87576, 92221, 96865, 101511, 106155, 110798, 115442, 120087, 124731, 129375, 134020, 138664, 143309, 147953, 152596, 157241, 161886, 166530, 171173, 175818, 180463, 185108, 189751, 194395, 199040, 203685, 208329, 212972, 217617, 222261, 226907, 231550, 236194, 240839, 245483, 250127, 254770, 259416, 264060, 268705, 273348, 277992, 282638, 287282, 291926, 296569, 301214, 305859, 310504, 315147, 319791, 324436, 329080, 333725, 338369, 343013, 347657, 352302, 356945, 361590, 366235]]
+    data1 = [ int(i*1.1959*1.085*1.051) for i in [0,45777, 50423, 55067, 59712, 64356, 68999, 73643, 78289, 82933, 87576, 92221, 96865, 101511, 106155, 110798, 115442, 120087, 124731, 129375, 134020, 138664, 143309, 147953, 152596, 157241, 161886, 166530, 171173, 175818, 180463, 185108, 189751, 194395, 199040, 203685, 208329, 212972, 217617, 222261, 226907, 231550, 236194, 240839, 245483, 250127, 254770, 259416, 264060, 268705, 273348, 277992, 282638, 287282, 291926, 296569, 301214, 305859, 310504, 315147, 319791, 324436, 329080, 333725, 338369, 343013, 347657, 352302, 356945, 361590, 366235]]
     攻击次数2 = 1
 
     CD = 40
@@ -426,11 +426,11 @@ class 技能18(职业主动技能):
     基础等级 = 13
     #基础 = (1872 - 190)*26 +12016 - 1219
     #成长 = 190 *26 + 1219
-    data0 = [ int(i*1.203*1.085) for i in [0,1873, 2063, 2252, 2443, 2633, 2823, 3013, 3204, 3393, 3583, 3774, 3963, 4154, 4344, 4533, 4724, 4914, 5104, 5294, 5485, 5674, 5864, 6055, 6244, 6435, 6625, 6814, 7005, 7195, 7385, 7575, 7766, 7955, 8145, 8336, 8528, 8717, 8907, 9098, 9287, 9478, 9668, 9857, 10048, 10238, 10428, 10618, 10809, 10998, 11188, 11379, 11568, 11759, 11949, 12138, 12329, 12519, 12709, 12899, 13090, 13279, 13469, 13660, 13849, 14040, 14230, 14419, 14610, 14800, 14990]]
+    data0 = [ int(i*1.203*1.085*1.052) for i in [0,1873, 2063, 2252, 2443, 2633, 2823, 3013, 3204, 3393, 3583, 3774, 3963, 4154, 4344, 4533, 4724, 4914, 5104, 5294, 5485, 5674, 5864, 6055, 6244, 6435, 6625, 6814, 7005, 7195, 7385, 7575, 7766, 7955, 8145, 8336, 8528, 8717, 8907, 9098, 9287, 9478, 9668, 9857, 10048, 10238, 10428, 10618, 10809, 10998, 11188, 11379, 11568, 11759, 11949, 12138, 12329, 12519, 12709, 12899, 13090, 13279, 13469, 13660, 13849, 14040, 14230, 14419, 14610, 14800, 14990]]
     攻击次数 = 1
-    data1 = [ int(i*1.203*1.085) for i in [0,1873, 2063, 2252, 2443, 2633, 2823, 3013, 3204, 3393, 3583, 3774, 3963, 4154, 4344, 4533, 4724, 4914, 5104, 5294, 5485, 5674, 5864, 6055, 6244, 6435, 6625, 6814, 7005, 7195, 7385, 7575, 7766, 7955, 8145, 8336, 8528, 8717, 8907, 9098, 9287, 9478, 9668, 9857, 10048, 10238, 10428, 10618, 10809, 10998, 11188, 11379, 11568, 11759, 11949, 12138, 12329, 12519, 12709, 12899, 13090, 13279, 13469, 13660, 13849, 14040, 14230, 14419, 14610, 14800, 14990]]
+    data1 = [ int(i*1.203*1.085*1.052) for i in [0,1873, 2063, 2252, 2443, 2633, 2823, 3013, 3204, 3393, 3583, 3774, 3963, 4154, 4344, 4533, 4724, 4914, 5104, 5294, 5485, 5674, 5864, 6055, 6244, 6435, 6625, 6814, 7005, 7195, 7385, 7575, 7766, 7955, 8145, 8336, 8528, 8717, 8907, 9098, 9287, 9478, 9668, 9857, 10048, 10238, 10428, 10618, 10809, 10998, 11188, 11379, 11568, 11759, 11949, 12138, 12329, 12519, 12709, 12899, 13090, 13279, 13469, 13660, 13849, 14040, 14230, 14419, 14610, 14800, 14990]]
     攻击次数2 = 25
-    data2 = [ int(i*1.203*1.085) for i in [0,12021, 13242, 14460, 15681, 16900, 18119, 19340, 20559, 21779, 22998, 24219, 25437, 26658, 27878, 29096, 30317, 31536, 32756, 33975, 35196, 36415, 37635, 38855, 40073, 41294, 42512, 43733, 44953, 46173, 47392, 48612, 49832, 51050, 52271, 53491, 54710, 55930, 57149, 58369, 59590, 60809, 62028, 63248, 64468, 65687, 66907, 68128, 69346, 70567, 71785, 73005, 74225, 75445, 76665, 77884, 79105, 80323, 81544, 82762, 83982, 85203, 86422, 87642, 88861, 90082, 91300, 92521, 93741, 94959, 96180]]
+    data2 = [ int(i*1.203*1.085*1.052) for i in [0,12021, 13242, 14460, 15681, 16900, 18119, 19340, 20559, 21779, 22998, 24219, 25437, 26658, 27878, 29096, 30317, 31536, 32756, 33975, 35196, 36415, 37635, 38855, 40073, 41294, 42512, 43733, 44953, 46173, 47392, 48612, 49832, 51050, 52271, 53491, 54710, 55930, 57149, 58369, 59590, 60809, 62028, 63248, 64468, 65687, 66907, 68128, 69346, 70567, 71785, 73005, 74225, 75445, 76665, 77884, 79105, 80323, 81544, 82762, 83982, 85203, 86422, 87642, 88861, 90082, 91300, 92521, 93741, 94959, 96180]]
     攻击次数3= 1
     CD = 45
