+++ Characters/重霄·弹药专家·男.py
@@ -41,8 +41,8 @@ class 技能0(被动技能):
     所在等级 = 15
     等级上限 = 20
     基础等级 = 10
-    关联技能 = ['交叉射击', '聚合弹', '凝固汽油弹', '轻火力速射']
-    关联技能2 = ['爆裂弹']
+    关联技能 = ['交叉射击', '聚合弹', '凝固汽油弹', '高压电磁弹']
+    关联技能2 = ['爆裂弹','尼尔狙击']
 
     def 加成倍率(self, 武器类型):
         if self.等级 == 0:
@@ -63,7 +63,7 @@ class 技能1(被动技能):
     等级上限 = 20
     基础等级 = 10
     冷却关联技能 = ['交叉射击', 'M18阔剑地雷', 'C4飞弹', 'G18冰冻手雷', 'G35感电手雷', 'G61重力手雷',
-              'G38ARG智能手雷', '爆裂弹', '聚合弹', '重火力支援', '镭射狙击', '凝固汽油弹', '轻火力速射']
+              'G38ARG智能手雷', '爆裂弹', '聚合弹', '重火力支援', '尼尔狙击', '凝固汽油弹', '高压电磁弹']
 
     def 加成倍率(self, 武器类型):
         return (1.1 + (self.等级 - 10) * 0.02) if self.等级 >= 10 else (1 + self.等级 * 0.01)
@@ -103,7 +103,7 @@ class 技能3(职业主动技能):
     等级精通 = 50
     # 基础 = 641.63
     # 成长 = 72.4
-    data0 = [int(i*1.1213) for i in [0, 714, 786, 859, 931, 1003, 1076, 1148, 1221, 1293, 1365, 1438, 1510, 1583, 1655, 1728, 1800, 1872, 1945, 2017, 2090, 2162, 2235, 2307, 2379, 2452, 2524, 2597, 2669, 2742, 2814, 2886, 2959, 3031, 3104, 3176, 3248, 3321, 3393, 3466, 3538, 3611, 3683, 3755, 3828, 3900, 3973, 4045, 4118, 4190, 4262, 4335, 4407, 4480, 4552, 4624, 4697, 4769, 4842, 4914, 4987, 5059, 5131, 5204, 5276, 5349, 5421, 5494, 5566, 5638, 5711]]
+    data0 = [int(i*1.1213*1.054) for i in [0, 714, 786, 859, 931, 1003, 1076, 1148, 1221, 1293, 1365, 1438, 1510, 1583, 1655, 1728, 1800, 1872, 1945, 2017, 2090, 2162, 2235, 2307, 2379, 2452, 2524, 2597, 2669, 2742, 2814, 2886, 2959, 3031, 3104, 3176, 3248, 3321, 3393, 3466, 3538, 3611, 3683, 3755, 3828, 3900, 3973, 4045, 4118, 4190, 4262, 4335, 4407, 4480, 4552, 4624, 4697, 4769, 4842, 4914, 4987, 5059, 5131, 5204, 5276, 5349, 5421, 5494, 5566, 5638, 5711]]
     攻击次数 = 3
     CD = 6.0
     TP成长 = 0.10
@@ -117,13 +117,11 @@ class 技能4(职业主动技能):
     等级上限 = 60
     学习间隔 = 2
     等级精通 = 50
-    data0 = [int(i*1.1514) for i in [0,540,595,650,705,760,814,869,924,979,1034,1089,1144,1198,1253,1308,1363,1418,1473,1527,1582,1637,1692,1747,1802,1856,1911,1966,2021,2076,2131,2185,2240,2295,2350,2405,2460,2515,2569,2624,2679,2734,2789,2844,2898,2953,3008,3063,3118,3173,3227,3282,3337,3392,3447,3502,3557,3611,3666,3721,3776]]
-    data1 = [int(i*1.00) for i in [0, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14,
-           14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14,
-           14, 14, 14, 14, 14, 14, 14]]
-    攻击次数2 = 18
+    data0 = [int(i*1.1514*3.0945) for i in [0,540,595,650,705,760,814,869,924,979,1034,1089,1144,1198,1253,1308,1363,1418,1473,1527,1582,1637,1692,1747,1802,1856,1911,1966,2021,2076,2131,2185,2240,2295,2350,2405,2460,2515,2569,2624,2679,2734,2789,2844,2898,2953,3008,3063,3118,3173,3227,3282,3337,3392,3447,3502,3557,3611,3666,3721,3776]]
+    data1 = [(0 if i ==0 else int(191 + i*22))for i in range(0,60)]
+    攻击次数2 = 1
 
-    CD = 3.0
+    CD = 8.0
     TP成长 = 0.10
     TP基础 = 5
     TP上限 = 5
@@ -142,7 +140,7 @@ class 技能5(职业主动技能):
     等级上限 = 60
     学习间隔 = 2
     等级精通 = 50
-    data0 = [int(i*1.1833) for i in [0, 1417, 1561, 1704, 1848, 1992, 2136, 2279, 2423, 2567, 2711, 2854, 2998, 3142, 3286, 3429, 3573, 3717, 3861, 4004, 4148, 4292, 4436, 4579, 4723, 4867, 5011, 5154, 5298, 5442, 5586, 5729, 5873, 6017, 6161, 6304, 6448, 6592, 6736, 6879, 7023, 7167, 7311, 7454, 7598, 7742, 7886, 8029, 8173, 8317, 8461, 8604, 8748, 8892, 9036, 9179, 9323, 9467, 9611, 9754, 9898, 10042, 10186, 10329, 10473, 10617, 10761, 10904, 11048, 11192, 11336]]
+    data0 = [int(i*1.1833*1.052) for i in [0, 1417, 1561, 1704, 1848, 1992, 2136, 2279, 2423, 2567, 2711, 2854, 2998, 3142, 3286, 3429, 3573, 3717, 3861, 4004, 4148, 4292, 4436, 4579, 4723, 4867, 5011, 5154, 5298, 5442, 5586, 5729, 5873, 6017, 6161, 6304, 6448, 6592, 6736, 6879, 7023, 7167, 7311, 7454, 7598, 7742, 7886, 8029, 8173, 8317, 8461, 8604, 8748, 8892, 9036, 9179, 9323, 9467, 9611, 9754, 9898, 10042, 10186, 10329, 10473, 10617, 10761, 10904, 11048, 11192, 11336]]
     攻击次数 = 3
     技能施放时间 = 0.3
     CD = 10.0
@@ -160,6 +158,8 @@ class 技能6(职业主动技能):
     # 基础 = 622.125
     # 成长 = 105.525
     data0 = [int(i*1.1601) for i in [0, 662, 768, 873, 978, 1084, 1189, 1294, 1400, 1506, 1611, 1717, 1822, 1927, 2033, 2138, 2244, 2350, 2455, 2561, 2666]]
+    # 子弹的额外倍率，但是存在浮动4-6hit,0.9暂定
+    倍率 = 1.37*0.91
     CD = 5.0
 
 
@@ -169,8 +169,8 @@ class 技能7(职业主动技能):
     等级上限 = 60
     学习间隔 = 2
     等级精通 = 50
-    data0 = [int(i*1.1567) for i in [0, 675, 744, 812, 881, 949, 1018, 1086, 1155, 1223, 1292, 1360, 1429, 1497, 1566, 1634, 1703, 1772, 1840, 1909, 1977, 2046, 2114, 2183, 2251, 2320, 2388, 2457, 2525, 2594, 2662, 2731, 2799, 2868, 2936, 3005, 3073, 3142, 3210, 3279, 3347, 3416, 3485, 3553, 3622, 3690, 3759, 3827, 3896, 3964, 4033, 4101, 4170, 4238, 4307, 4375, 4444, 4512, 4581, 4649, 4718, 4786, 4855, 4923, 4992, 5060, 5129, 5197, 5266, 5335, 5403]]
-    CD = 4.0
+    data0 = [int(i*1.1567*3.175) for i in [0, 675, 744, 812, 881, 949, 1018, 1086, 1155, 1223, 1292, 1360, 1429, 1497, 1566, 1634, 1703, 1772, 1840, 1909, 1977, 2046, 2114, 2183, 2251, 2320, 2388, 2457, 2525, 2594, 2662, 2731, 2799, 2868, 2936, 3005, 3073, 3142, 3210, 3279, 3347, 3416, 3485, 3553, 3622, 3690, 3759, 3827, 3896, 3964, 4033, 4101, 4170, 4238, 4307, 4375, 4444, 4512, 4581, 4649, 4718, 4786, 4855, 4923, 4992, 5060, 5129, 5197, 5266, 5335, 5403]]
+    CD = 10.0
     TP成长 = 0.10
     TP基础 = 5
     TP上限 = 5
@@ -184,11 +184,22 @@ class 技能8(职业主动技能):
     等级精通 = 50
     # 基础 = 8206.58
     # 成长 = 926.53341
-    data0 = [int(i*1.1847) for i in [0, 9129, 10056, 10982, 11908, 12834, 13760, 14687, 15613, 16539, 17465, 18392, 19318, 20244, 21170, 22096, 23023, 23949, 24875, 25801, 26728, 27654, 28580, 29506, 30432, 31359, 32285, 33211, 34137, 35064, 35990, 36916, 37842, 38768, 39695, 40621, 41547, 42473, 43399, 44326, 45252, 46178, 47104, 48031, 48957, 49883, 50809, 51735, 52662, 53588, 54514, 55440, 56367, 57293, 58219, 59145, 60071, 60998, 61924, 62850, 63776, 64703, 65629, 66555, 67481, 68407, 69334, 70260, 71186, 72112, 73038]]
+    data0 = [int(i*1.1847*1.053) for i in [0, 9129, 10056, 10982, 11908, 12834, 13760, 14687, 15613, 16539, 17465, 18392, 19318, 20244, 21170, 22096, 23023, 23949, 24875, 25801, 26728, 27654, 28580, 29506, 30432, 31359, 32285, 33211, 34137, 35064, 35990, 36916, 37842, 38768, 39695, 40621, 41547, 42473, 43399, 44326, 45252, 46178, 47104, 48031, 48957, 49883, 50809, 51735, 52662, 53588, 54514, 55440, 56367, 57293, 58219, 59145, 60071, 60998, 61924, 62850, 63776, 64703, 65629, 66555, 67481, 68407, 69334, 70260, 71186, 72112, 73038]]
     CD = 18.0
+    data1 = [(0 if i ==0 else int(4178+(i-1)*424)) for i in range(0,60)]
+    攻击次数2 = 0
+    data2 = [(0 if i ==0 else int(3249+(i-1)*330)) for i in range(0,60)]
+    攻击次数3 = 0
     TP成长 = 0.10
     TP基础 = 5
     TP上限 = 5
+
+    def 等效百分比(self, 武器类型):
+        if 武器类型 == '手弩':
+            self.攻击次数 = 0
+            self.攻击次数2 = 1
+            self.攻击次数3 = 3
+        return super().等效百分比(武器类型)
 
 
 class 技能9(职业主动技能):
@@ -199,7 +210,7 @@ class 技能9(职业主动技能):
     等级精通 = 50
     # 基础 = 7122.49875
     # 成长 = 804.25125
-    data0 = [int(i*1.1598) for i in [0, 7927, 8732, 9536, 10340, 11144, 11949, 12753, 13557, 14361, 15166, 15970, 16774, 17579, 18383, 19187, 19991, 20796, 21600, 22404, 23208, 24013, 24817, 25621, 26426, 27230, 28034, 28838, 29643, 30447, 31251, 32055, 32860, 33664, 34468, 35272, 36077, 36881, 37685, 38490, 39294, 40098, 40902, 41707, 42511, 43315, 44119, 44924, 45728, 46532, 47337, 48141, 48945, 49749, 50554, 51358, 52162, 52966, 53771, 54575, 55379, 56183, 56988, 57792, 58596, 59401, 60205, 61009, 61813, 62618, 63422]]
+    data0 = [int(i*1.1598*1.051) for i in [0, 7927, 8732, 9536, 10340, 11144, 11949, 12753, 13557, 14361, 15166, 15970, 16774, 17579, 18383, 19187, 19991, 20796, 21600, 22404, 23208, 24013, 24817, 25621, 26426, 27230, 28034, 28838, 29643, 30447, 31251, 32055, 32860, 33664, 34468, 35272, 36077, 36881, 37685, 38490, 39294, 40098, 40902, 41707, 42511, 43315, 44119, 44924, 45728, 46532, 47337, 48141, 48945, 49749, 50554, 51358, 52162, 52966, 53771, 54575, 55379, 56183, 56988, 57792, 58596, 59401, 60205, 61009, 61813, 62618, 63422]]
     CD = 20.0
     TP成长 = 0.10
     TP基础 = 5
@@ -241,13 +252,18 @@ class 技能10(职业主动技能):
             self.CD *= 0.94
 
 
-class 技能11(职业主动技能):
-    名称 = '镭射狙击'
+class 技能11(主动技能):
+    名称 = '尼尔狙击'
+    脱手 = 0
+    技能施放时间 = 4
     所在等级 = 45
     等级上限 = 60
     学习间隔 = 2
     等级精通 = 50
-    data0 = [int(i*1.1693) for i in [0, 3360, 3701, 4042, 4383, 4724, 5065, 5406, 5747, 6088, 6429, 6770, 7110, 7451, 7792, 8133, 8474, 8815, 9156, 9497, 9838, 10179, 10520, 10861, 11202, 11542, 11883, 12224, 12565, 12906, 13247, 13588, 13929, 14270, 14611, 14952, 15293, 15633, 15974, 16315, 16656, 16997, 17338, 17679, 18020, 18361, 18702, 19043, 19384, 19724, 20065, 20406, 20747, 21088, 21429, 21770, 22111, 22452, 22793, 23134, 23475, 23815, 24156, 24497, 24838, 25179, 25520, 25861, 26202, 26543, 26884]]
+    # data0 = [int(i*1.1693*1.51798) for i in [0, 3360, 3701, 4042, 4383, 4724, 5065, 5406, 5747, 6088, 6429, 6770, 7110, 7451, 7792, 8133, 8474, 8815, 9156, 9497, 9838, 10179, 10520, 10861, 11202, 11542, 11883, 12224, 12565, 12906, 13247, 13588, 13929, 14270, 14611, 14952, 15293, 15633, 15974, 16315, 16656, 16997, 17338, 17679, 18020, 18361, 18702, 19043, 19384, 19724, 20065, 20406, 20747, 21088, 21429, 21770, 22111, 22452, 22793, 23134, 23475, 23815, 24156, 24497, 24838, 25179, 25520, 25861, 26202, 26543, 26884]]
+    基础 = 5966 - 605
+    成长 = 605
+    # 每次会附加一发子弹，不享受70和一觉被动
     攻击次数 = 5
     CD = 45.0
     TP成长 = 0.10
@@ -259,10 +275,13 @@ class 技能11(职业主动技能):
 
     def 装备护石(self, x):
         if x == 0:
-            self.倍率 *= 1.28
-            # self.倍率 *= 1.18
+            self.技能施放时间 = 2
+            self.倍率 *= 0.75*1.06
+            self.攻击次数 += 3
         elif x == 1:
-            self.倍率 *= 1.36
+            self.技能施放时间 = 2
+            self.倍率 *= 0.75*1.09
+            self.攻击次数 += 3
 
 
 class 技能12(被动技能):
@@ -290,13 +309,13 @@ class 技能13(职业主动技能):
     等级上限 = 40
     学习间隔 = 5
     等级精通 = 30
-    data0 = [int(i*1.2545) for i in [0, 1789, 2204, 2619, 3034, 3448, 3863, 4278, 4693, 5108, 5523, 5938, 6353, 6768, 7183, 7598, 8013, 8427, 8842, 9257, 9672, 10087, 10502, 10917, 11332, 11747, 12162, 12577, 12991, 13406, 13821, 14236, 14651, 15066, 15481, 15896, 16311, 16726, 17141, 17556, 17970, 18385, 18800, 19215, 19630, 20045, 20460, 20875, 21290, 21705, 22120, 22535, 22949, 23364, 23779, 24194, 24609, 25024, 25439, 25854, 26269, 26684, 27099, 27513, 27928, 28343, 28758, 29173, 29588, 30003, 30418]]
+    data0 = [int(i*1.2545*1.053) for i in [0, 1789, 2204, 2619, 3034, 3448, 3863, 4278, 4693, 5108, 5523, 5938, 6353, 6768, 7183, 7598, 8013, 8427, 8842, 9257, 9672, 10087, 10502, 10917, 11332, 11747, 12162, 12577, 12991, 13406, 13821, 14236, 14651, 15066, 15481, 15896, 16311, 16726, 17141, 17556, 17970, 18385, 18800, 19215, 19630, 20045, 20460, 20875, 21290, 21705, 22120, 22535, 22949, 23364, 23779, 24194, 24609, 25024, 25439, 25854, 26269, 26684, 27099, 27513, 27928, 28343, 28758, 29173, 29588, 30003, 30418]]
     # 基础 = 1373.97702
     # 成长 = 415.33411
     攻击次数 = 16
     # 基础2 = 242.389
     # 成长2 = 73.91189
-    data1 = [int(i*1.2542) for i in [0, 318, 391, 465, 539, 613, 686, 760, 834, 908, 981, 1055, 1129, 1203, 1277, 1350, 1424, 1498, 1572, 1645, 1719, 1793, 1867, 1940, 2014, 2088, 2162, 2235, 2309, 2383, 2457, 2530, 2604, 2678, 2752, 2826, 2899, 2973, 3047, 3121, 3194, 3268, 3342, 3416, 3489, 3563, 3637, 3711, 3784, 3858, 3932, 4006, 4079, 4153, 4227, 4301, 4375, 4448, 4522, 4596, 4670, 4743, 4817, 4891, 4965, 5038, 5112, 5186, 5260, 5333, 5407]]
+    data1 = [int(i*1.2542*1.053) for i in [0, 318, 391, 465, 539, 613, 686, 760, 834, 908, 981, 1055, 1129, 1203, 1277, 1350, 1424, 1498, 1572, 1645, 1719, 1793, 1867, 1940, 2014, 2088, 2162, 2235, 2309, 2383, 2457, 2530, 2604, 2678, 2752, 2826, 2899, 2973, 3047, 3121, 3194, 3268, 3342, 3416, 3489, 3563, 3637, 3711, 3784, 3858, 3932, 4006, 4079, 4153, 4227, 4301, 4375, 4448, 4522, 4596, 4670, 4743, 4817, 4891, 4965, 5038, 5112, 5186, 5260, 5333, 5407]]
     # 攻击次数2 = 90 10hit左右无法打到
     攻击次数2 = 80
     CD = 145.0
@@ -332,16 +351,19 @@ class 技能14(职业主动技能):
             self.攻击次数2 = 15
 
 
-class 技能15(职业主动技能):
-    名称 = '轻火力速射'
-    脱手 = 0
+class 技能15(主动技能):
+    名称 = '高压电磁弹'
+    脱手 = 1
     技能施放时间 = 1
     所在等级 = 70
     等级上限 = 40
     学习间隔 = 2
     等级精通 = 30
-    data0 = [int(i*1.1814) for i in [0, 1229, 1353, 1478, 1602, 1727, 1852, 1976, 2101, 2226, 2350, 2475, 2599, 2724, 2849, 2973, 3098, 3223, 3347, 3472, 3597, 3721, 3846, 3970, 4095, 4220, 4344, 4469, 4594, 4718, 4843, 4968, 5092, 5217, 5341, 5466, 5591, 5715, 5840, 5965, 6089, 6214, 6339, 6463, 6588, 6712, 6837, 6962, 7086, 7211, 7336, 7460, 7585, 7710, 7834, 7959, 8083, 8208, 8333, 8457, 8582, 8707, 8831, 8956, 9080, 9205, 9330, 9454, 9579, 9704, 9828]]
-    攻击次数 = 17
+    # data0 = [int(i*1.1814*1.0327) for i in [0, 1229, 1353, 1478, 1602, 1727, 1852, 1976, 2101, 2226, 2350, 2475, 2599, 2724, 2849, 2973, 3098, 3223, 3347, 3472, 3597, 3721, 3846, 3970, 4095, 4220, 4344, 4469, 4594, 4718, 4843, 4968, 5092, 5217, 5341, 5466, 5591, 5715, 5840, 5965, 6089, 6214, 6339, 6463, 6588, 6712, 6837, 6962, 7086, 7211, 7336, 7460, 7585, 7710, 7834, 7959, 8083, 8208, 8333, 8457, 8582, 8707, 8831, 8956, 9080, 9205, 9330, 9454, 9579, 9704, 9828]]
+    # 攻击次数 = 17
+    基础 = 25491 - 2587
+    成长 = 2587
+
     CD = 30.0
     TP成长 = 0.10
     TP基础 = 5
@@ -354,17 +376,15 @@ class 技能15(职业主动技能):
         if x == 0:
             # self.基础2 = self.基础*0.22
             # self.成长2 = self.成长*0.22
-            self.技能施放时间 = 2
             # self.攻击次数2 = 17
             self.CD *= 0.95
-            self.倍率 *= 1.22
+            self.倍率 *= 1.25
         elif x == 1:
             # self.基础2 = self.基础*0.36
             # self.成长2 = self.成长*0.36
-            self.技能施放时间 = 2
             # self.攻击次数2 = 17
             self.CD *= 0.95
-            self.倍率 *= 1.32
+            self.倍率 *= 1.33
 
 
 class 技能16(被动技能):
@@ -407,11 +427,11 @@ class 技能18(职业主动技能):
     等级精通 = 30
     # 基础 = 13127.83*1.127
     # 成长 = 1482.17*1.127
-    data0 = [int(i*1.1534) for i in [0, 16468, 18139, 19809, 21480, 23151, 24821, 26492, 28163, 29833, 31504, 33175, 34846, 36516, 38187, 39858, 41528, 43199, 44870, 46541, 48211, 49882, 51553, 53223, 54894, 56565, 58235, 59906, 61577, 63248, 64918, 66589, 68260, 69930, 71601, 73272, 74942, 76613, 78284, 79955, 81625, 83296, 84967, 86637, 88308, 89979, 91649, 93320, 94991, 96662, 98332, 100003, 101674, 103344, 105015, 106686, 108357, 110027, 111698, 113369, 115039, 116710, 118381, 120051, 121722, 123393, 125064, 126734, 128405, 130076, 131746]]
+    data0 = [int(i*1.1534*1.052) for i in [0, 16468, 18139, 19809, 21480, 23151, 24821, 26492, 28163, 29833, 31504, 33175, 34846, 36516, 38187, 39858, 41528, 43199, 44870, 46541, 48211, 49882, 51553, 53223, 54894, 56565, 58235, 59906, 61577, 63248, 64918, 66589, 68260, 69930, 71601, 73272, 74942, 76613, 78284, 79955, 81625, 83296, 84967, 86637, 88308, 89979, 91649, 93320, 94991, 96662, 98332, 100003, 101674, 103344, 105015, 106686, 108357, 110027, 111698, 113369, 115039, 116710, 118381, 120051, 121722, 123393, 125064, 126734, 128405, 130076, 131746]]
     攻击次数 = 1
     # 基础2 = 3063.17*1.127
     # 成长2 = 345.83*1.127
-    data1 = [int(i*1.1534) for i in [0, 3842, 4232, 4622, 5012, 5401, 5791, 6181, 6571, 6961, 7351, 7740, 8130, 8520, 8910, 9300, 9690, 10079, 10469, 10859, 11249, 11639, 12029, 12418, 12808, 13198, 13588, 13978, 14368, 14757, 15147, 15537, 15927, 16317, 16707, 17096, 17486, 17876, 18266, 18656, 19046, 19435, 19825, 20215, 20605, 20995, 21384, 21774, 22164, 22554, 22944, 23334, 23723, 24113, 24503, 24893, 25283, 25673, 26062, 26452, 26842, 27232, 27622, 28012, 28401, 28791, 29181, 29571, 29961, 30351, 30740]]
+    data1 = [int(i*1.1534*1.052) for i in [0, 3842, 4232, 4622, 5012, 5401, 5791, 6181, 6571, 6961, 7351, 7740, 8130, 8520, 8910, 9300, 9690, 10079, 10469, 10859, 11249, 11639, 12029, 12418, 12808, 13198, 13588, 13978, 14368, 14757, 15147, 15537, 15927, 16317, 16707, 17096, 17486, 17876, 18266, 18656, 19046, 19435, 19825, 20215, 20605, 20995, 21384, 21774, 22164, 22554, 22944, 23334, 23723, 24113, 24503, 24893, 25283, 25673, 26062, 26452, 26842, 27232, 27622, 28012, 28401, 28791, 29181, 29571, 29961, 30351, 30740]]
     攻击次数2 = 10
     CD = 40.0
     是否有护石 = 1
@@ -429,8 +449,8 @@ class 技能19(职业主动技能):
     等级上限 = 40
     学习间隔 = 5
     等级精通 = 30
-    data0 = [int(i*1.24) for i in [0, 71384, 87937, 104490, 121043, 137596, 154149, 170702, 187254, 203807, 220360, 236913, 253466, 270019, 286572, 303125, 319678, 336231, 352784, 369337, 385889, 402442, 418995, 435548, 452101, 468654, 485207, 501760, 518313, 534866, 551419, 567972, 584525, 601077, 617630, 634183, 650736, 667289, 683842, 700395, 716948, 733501, 750054, 766607, 783160, 799713, 816265, 832818, 849371, 865924, 882477, 899030, 915583, 932136, 948689, 965242, 981795, 998348, 1014900, 1031453, 1048006, 1064559, 1081112, 1097665, 1114218, 1130771, 1147324, 1163877, 1180430, 1196983, 1213536]]
-    data1 = [int(i*1.24) for i in [0, 2039, 2512, 2985, 3458, 3931, 4404, 4877, 5350, 5823, 6296, 6768, 7241, 7714, 8187, 8660, 9133, 9606, 10079, 10552, 11025, 11498, 11971, 12444, 12917, 13390, 13863, 14336, 14808, 15281, 15754, 16227, 16700, 17173, 17646, 18119, 18592, 19065, 19538, 20011, 20484, 20957, 21430, 21903, 22376, 22848, 23321, 23794, 24267, 24740, 25213, 25686, 26159, 26632, 27105, 27578, 28051, 28524, 28997, 29470, 29943, 30415, 30888, 31361, 31834, 32307, 32780, 33253, 33726, 34199, 34672]]
+    data0 = [int(i*1.24*1.063) for i in [0, 71384, 87937, 104490, 121043, 137596, 154149, 170702, 187254, 203807, 220360, 236913, 253466, 270019, 286572, 303125, 319678, 336231, 352784, 369337, 385889, 402442, 418995, 435548, 452101, 468654, 485207, 501760, 518313, 534866, 551419, 567972, 584525, 601077, 617630, 634183, 650736, 667289, 683842, 700395, 716948, 733501, 750054, 766607, 783160, 799713, 816265, 832818, 849371, 865924, 882477, 899030, 915583, 932136, 948689, 965242, 981795, 998348, 1014900, 1031453, 1048006, 1064559, 1081112, 1097665, 1114218, 1130771, 1147324, 1163877, 1180430, 1196983, 1213536]]
+    data1 = [int(i*1.24*1.063) for i in [0, 2039, 2512, 2985, 3458, 3931, 4404, 4877, 5350, 5823, 6296, 6768, 7241, 7714, 8187, 8660, 9133, 9606, 10079, 10552, 11025, 11498, 11971, 12444, 12917, 13390, 13863, 14336, 14808, 15281, 15754, 16227, 16700, 17173, 17646, 18119, 18592, 19065, 19538, 20011, 20484, 20957, 21430, 21903, 22376, 22848, 23321, 23794, 24267, 24740, 25213, 25686, 26159, 26632, 27105, 27578, 28051, 28524, 28997, 29470, 29943, 30415, 30888, 31361, 31834, 32307, 32780, 33253, 33726, 34199, 34672]]
     攻击次数 = 1
     攻击次数2 = 15
     脱手 = 0
@@ -593,42 +613,44 @@ class 重霄·弹药专家·男角色属性(角色属性):
         技能消耗时间 = 0.0
         爆裂弹间隔 = 0.115
         每轮射击次数 = 0.5 * \
-            (16 + floor(0.5 * (min(self.技能栏[self.技能序号['弹夹改装']].等级, 20)-10)))
+            (17 + floor(0.5 * (min(self.技能栏[self.技能序号['弹夹改装']].等级, 20)-10)))
 
         if (self.武器类型 != '手弩'):
             爆裂弹间隔 = 0.14
             每轮射击次数 = 0.5 * \
-                (9 + floor(0.5 * (min(self.技能栏[self.技能序号['弹夹改装']].等级, 20)-9)))
+                (10 + floor(0.5 * (min(self.技能栏[self.技能序号['弹夹改装']].等级, 20)-9)))
+
+        狙击附加子弹 = 0
 
         for i in self.技能栏:
             if i.是否有伤害 == 1:
                 if i.名称 == '爆裂弹':
                     技能释放次数.append(0)
                 else:
+                    count = 0
                     if '/CD' in self.次数输入[self.技能序号[i.名称]]:
-                        技能释放次数.append(
-                            int((self.时间输入) / (i.等效CD(self.武器类型, self.类型)+i.技能施放时间) + 1 + i.基础释放次数))
+                        count = int((self.时间输入) / (i.等效CD(self.武器类型, self.类型)+i.技能施放时间) + 1 + i.基础释放次数)
                         if i.脱手 == 1:
-                            技能消耗时间 += int((self.时间输入) / (i.等效CD(self.武器类型,
-                                          self.类型) + i.技能施放时间) + 1 + i.基础释放次数) * 0.2
+                            技能消耗时间 += count * 0.2
                         else:
-                            技能消耗时间 += int((self.时间输入) / (i.等效CD(self.武器类型,
-                                          self.类型) + i.技能施放时间) + 1 + i.基础释放次数) * i.技能施放时间
+                            技能消耗时间 += count * i.技能施放时间
                     elif self.次数输入[self.技能序号[i.名称]] != '0':
-                        技能释放次数.append(
-                            round(float(self.次数输入[self.技能序号[i.名称]]), 2))
+                        count = round(float(self.次数输入[self.技能序号[i.名称]]), 2)
                     else:
-                        技能释放次数.append(0)
+                        count = 0
+                    技能释放次数.append(count)
+                    if i.名称 == '尼尔狙击':
+                        狙击附加子弹 = count * i.攻击次数
+
             else:
                 技能释放次数.append(0)
 
         if '/CD' in self.次数输入[self.技能序号['爆裂弹']]:
             每轮射击时间 = 每轮射击次数 * 爆裂弹间隔
             射击轮次 = int((self.时间输入-技能消耗时间)/每轮射击时间)
-            技能释放次数[self.技能序号['爆裂弹']] = int(floor((self.时间输入-技能消耗时间-射击轮次*0.2)/爆裂弹间隔))
+            技能释放次数[self.技能序号['爆裂弹']] = int(floor((self.时间输入-技能消耗时间-射击轮次*0.2)/爆裂弹间隔)) + round(狙击附加子弹/1.3,2)
         else:
             if self.次数输入[self.技能序号['爆裂弹']] != '0':
-                print()
                 技能释放次数[self.技能序号['爆裂弹']] = round(
                     float(self.次数输入[self.技能序号['爆裂弹']]), 2)*每轮射击次数
             else:
