+++ ResourceFiles/Config/adventure_info.json
@@ -21,7 +21,7 @@
     "三次觉醒": "极诣·鬼泣",
     "显示名称": "极诣·鬼泣",
     "类名": "极诣·鬼泣",
-    "类名2": "TYF.极诣·鬼泣",
+    "类名2": "无",
     "序号": "2",
     "作者": "真奥",
     "时间": "2020/6/19",
@@ -147,7 +147,7 @@
     "三次觉醒": "归元·气功师",
     "显示名称": "归元·气功师",
     "类名": "归元·气功师·男",
-    "类名2": "TYF.归元·气功师·男",
+    "类名2": "无",
     "序号": "11",
     "作者": "球",
     "时间": "2020/6/26",
@@ -329,7 +329,7 @@
     "三次觉醒": "重霄·弹药专家",
     "显示名称": "重霄·弹药专家",
     "类名": "重霄·弹药专家·男",
-    "类名2": "TYF.重霄·弹药专家·男",
+    "类名2": "无",
     "序号": "24",
     "作者": "小一",
     "时间": "2020/6/8",
@@ -455,7 +455,7 @@
     "三次觉醒": "知源·血法师",
     "显示名称": "知源·血法师",
     "类名": "知源·血法师",
-    "类名2": "TYF.知源·血法师",
+    "类名2": "无",
     "序号": "33",
     "作者": "莫方、阿武",
     "时间": "2020/7/11",
@@ -511,7 +511,7 @@
     "三次觉醒": "知源·召唤师",
     "显示名称": "知源·召唤师",
     "类名": "知源·召唤师",
-    "类名2": "TYF.知源·召唤师",
+    "类名2": "无",
     "序号": "37",
     "作者": "Lizzy荔枝贝壳喵",
     "时间": "2020/7/4",
@@ -655,7 +655,7 @@
     "三次觉醒": "神启·异端审判者",
     "显示名称": "神启·异端审判者",
     "类名": "神启·异端审判者",
-    "类名2": "TYF.神启·异端审判者",
+    "类名2": "无",
     "序号": "47",
     "作者": "熬夜砍树、乐于驴",
     "时间": "2020/7/6",
@@ -683,7 +683,7 @@
     "三次觉醒": "神启·诱魔者",
     "显示名称": "神启·诱魔者",
     "类名": "神启·诱魔者",
-    "类名2": "TYF.神启·诱魔者",
+    "类名2": "无",
     "序号": "49",
     "作者": "千景",
     "时间": "2020/6/21",
@@ -809,7 +809,7 @@
     "三次觉醒": "皓曦·帕拉丁",
     "显示名称": "皓曦·帕拉丁",
     "类名": "皓曦·帕拉丁",
-    "类名2": "TYF.皓曦·帕拉丁",
+    "类名2": "无",
     "序号": "58",
     "作者": "疯",
     "时间": "2020/6/13",
@@ -991,7 +991,7 @@
     "三次觉醒": "",
     "显示名称": "知源·缔造者",
     "类名": "知源·缔造者",
-    "类名2": "TYF.知源·缔造者",
+    "类名2": "无",
     "序号": "71",
     "作者": "时雨、🐎、纸飞机",
     "时间": "2020/7/4",
