+++ ResourceFiles/Config/release_version.json
@@ -1,5 +1,5 @@
 {
-  "version": "2022-02-14",
+  "version": "Lv100-LastVersion",
   "ShowChangeLog": true,
   "agreement": false
 }
